Print Community Object

3 function for print the community object

Usage

# S3 method for class 'community_object'
print(x, ...)

Arguments

x

the object created from the create_community_object() function.

...

other parameters that are included in the print function.

Examples

data <-
   import_all_data(peak_table =
                   mums2::mums2_example("full_mix_peak_table_small.csv"),
                   meta_data =
                   mums2::mums2_example("full_mix_meta_data_small.csv"),
                   format = "Metaboscape")
#> If peak table has corrupted compound names they will be converted to
#>       utf-8 and if there are any commas, they will be converted to periods(.).

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())
#> ℹ Checking 1294 peaks for mispicked peaks.
#> ℹ Argument merge_peaks is: TRUE. Merging mispicked peaks with method sum.
#> ✔ 8 ions failed the mispicked filter, 1286 ions remain.
#> ℹ Parsing 1286 peaks for replicability across technical replicates.
#> ✔ 399 ions failed the cv_filter filter, 887 ions remain.
#> ℹ Parsing 887 peaks based on the sample group: Blanks.
#> ℹ Argument remove_ions is: TRUE.Removing peaks from Blanks.
#> ✔ 494 ions failed the Blanks filter, 393 ions remain.
#> ℹ Parsing 393 peaks for insource ions.
#> ✔ 43 ions failed the insource filter, 350 ions remain.

change_rt_to_seconds_or_minute(filtered_data, "minutes")
#> [1] "Changing rt values to minutes"
#> Key: <Compound, mz, kmd, RTINMINUTES>
#>      Compound        mz      kmd RTINMINUTES 50MeOH_10570 50MeOH_10571
#>        <char>     <num>    <num>       <num>        <num>        <num>
#>   1:     1000 749.51361 0.513606    3.622667            0            0
#>   2:     1002 734.46721 0.467206    3.725667            0            0
#>   3:     1003 749.51369 0.513686    3.730833            0            0
#>   4:     1006  42.03365 0.033646    3.802167            0            0
#>   5:     1007 137.07102 0.071016    3.807167            0            0
#>  ---                                                                  
#> 346:      989  83.04884 0.048836    3.307833            0            0
#> 347:      992 749.51424 0.514236    3.343167            0            0
#> 348:      996 716.45731 0.457306    3.447000            0            0
#> 349:      997 749.51417 0.514166    3.524000            0            0
#> 350:      999 558.36355 0.363546    3.556833            0            0
#>      50MeOH_10572 FullMix_10574 FullMix_10577 FullMix_10580    cor
#>             <num>         <num>         <num>         <num> <lgcl>
#>   1:            0     226277.39     187860.42     166603.72   TRUE
#>   2:            0    1285652.63    1244527.38    1222279.13   TRUE
#>   3:            0     325578.88     357518.06     316568.66   TRUE
#>   4:            0     270776.94     258514.84     251178.77   TRUE
#>   5:            0     116058.97     118762.07     102857.34   TRUE
#>  ---                                                              
#> 346:            0     198559.53     206690.77     217850.67   TRUE
#> 347:            0      86088.67      83416.42      65714.98   TRUE
#> 348:            0    1578886.38    1743622.13    1736709.88   TRUE
#> 349:            0     129910.10     137885.48     128256.25   TRUE
#> 350:            0     192658.81     218406.34     205511.38   TRUE

matched_data <- ms2_ms1_compare(mums2_example("full_mix_ms2_small.mgf"),
 filtered_data, 2, 6)
#> [1] "Reading: /home/runner/work/_temp/Library/mums2/extdata/full_mix_ms2_small.mgf ..."
#> Computing                                                    | 0%  ETA: -...Computing ■                                                  | 2%  ETA: ...Computing ■■                                                 | 4%  ETA: ...Computing ■■■                                                | 6%  ETA: ...Computing ■■■■                                               | 8%  ETA: ...Computing ■■■■■                                              | 10%  ETA: ...Computing ■■■■■■                                             | 12%  ETA: ...Computing ■■■■■■■                                            | 14%  ETA: ...Computing ■■■■■■■■                                           | 16%  ETA: ...Computing ■■■■■■■■■                                          | 18%  ETA: ...Computing ■■■■■■■■■■                                         | 20%  ETA: ...Computing ■■■■■■■■■■■                                        | 22%  ETA: ...Computing ■■■■■■■■■■■■                                       | 24%  ETA: ...Computing ■■■■■■■■■■■■■                                      | 26%  ETA: ...Computing ■■■■■■■■■■■■■■                                     | 28%  ETA: ...Computing ■■■■■■■■■■■■■■■                                    | 30%  ETA: ...Computing ■■■■■■■■■■■■■■■■                                   | 32%  ETA: ...Computing ■■■■■■■■■■■■■■■■■                                  | 34%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■                                 | 36%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■                                | 38%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■                               | 40%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■                              | 42%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■                             | 44%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■                            | 46%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■                           | 48%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■                          | 50%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■                         | 52%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■                        | 54%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■                       | 56%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                      | 58%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                     | 60%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                    | 62%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                   | 64%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                  | 66%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                 | 68%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                | 70%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■               | 72%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■              | 74%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■             | 76%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            | 78%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■           | 80%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■          | 82%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■         | 84%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■        | 86%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■       | 88%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■      | 90%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     | 92%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    | 94%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   | 96%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  | 98%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ | 100%  ETA: ...
#> [1] "214/350 peaks have an MS2 spectra."

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = modified_cosine_params(0.5), min_peaks = 0)
#> Computing                                                    | 0%  ETA: -...Computing ■                                                  | 2%  ETA: ...Computing ■■                                                 | 4%  ETA: ...Computing ■■■                                                | 6%  ETA: ...Computing ■■■■                                               | 8%  ETA: 10s ...Computing ■■■■■                                              | 10%  ETA: 8s ...Computing ■■■■■■                                             | 12%  ETA: 7s ...Computing ■■■■■■■                                            | 14%  ETA: 6s ...Computing ■■■■■■■■                                           | 16%  ETA: 5s ...Computing ■■■■■■■■■                                          | 18%  ETA: 4s ...Computing ■■■■■■■■■■                                         | 20%  ETA: 3s ...Computing ■■■■■■■■■■■                                        | 22%  ETA: 3s ...Computing ■■■■■■■■■■■■                                       | 24%  ETA: 3s ...Computing ■■■■■■■■■■■■■                                      | 26%  ETA: 2s ...Computing ■■■■■■■■■■■■■■                                     | 28%  ETA: 2s ...Computing ■■■■■■■■■■■■■■■                                    | 30%  ETA: 2s ...Computing ■■■■■■■■■■■■■■■■                                   | 32%  ETA: 2s ...Computing ■■■■■■■■■■■■■■■■■                                  | 34%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■                                 | 36%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■                                | 38%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■                               | 40%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■■                              | 42%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■■■                             | 44%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■■■■                            | 46%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■                           | 48%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■                          | 50%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■                         | 52%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■                        | 54%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■                       | 56%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                      | 58%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                     | 60%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                    | 62%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                   | 64%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                  | 66%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                 | 68%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                | 70%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■               | 72%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■              | 74%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■             | 76%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            | 78%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■           | 80%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■          | 82%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■         | 84%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■        | 86%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■       | 88%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■      | 90%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     | 92%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    | 94%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   | 96%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  | 98%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ | 100%  ETA: ...

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data, cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)
print(community_object)
#>                  omu1    omu2     omu3     omu4    omu5     omu6    omu7
#> FullMix_10577 2998098 1933658 19650.86 24121.65 1587891 593985.6 3480016
#> FullMix_10580 2843044 1858876 21177.87 23067.25 1561004 565654.8 3241200
#> FullMix_10574 2943366 1934672 20072.26 24968.87 1545093 566948.9 3159052
#> 50MeOH_10572        0       0     0.00     0.00       0      0.0       0
#> 50MeOH_10571        0       0     0.00     0.00       0      0.0       0
#> 50MeOH_10570        0       0     0.00     0.00       0      0.0       0
#>                   omu8      omu9   omu10   omu11   omu12    omu13    omu14
#> FullMix_10577 11169.21 788243.88 3960.08 6152.61 6146.53 19755.67 17726.77
#> FullMix_10580 10920.95 737723.38 4580.26 5918.13 5495.50 18521.56 17838.09
#> FullMix_10574  9493.06 746605.38 4331.92 5628.47 6060.95 18587.09 18966.64
#> 50MeOH_10572      0.00  28821.89    0.00    0.00    0.00     0.00     0.00
#> 50MeOH_10571      0.00  28947.73    0.00    0.00    0.00     0.00     0.00
#> 50MeOH_10570      0.00  26125.07    0.00    0.00    0.00     0.00     0.00
#>                  omu15    omu16     omu17    omu18    omu19    omu20    omu21
#> FullMix_10577 364303.6 220695.4  98912.84 33625.80 27668.39 271628.0 256334.0
#> FullMix_10580 355948.7 233698.2  97682.88 32395.65 29053.76 260073.2 245960.0
#> FullMix_10574 416666.7 194363.3 104566.88 33121.33 30538.18 289277.6 258184.6
#> 50MeOH_10572       0.0      0.0      0.00     0.00     0.00      0.0      0.0
#> 50MeOH_10571       0.0      0.0      0.00     0.00     0.00      0.0      0.0
#> 50MeOH_10570       0.0      0.0      0.00     0.00     0.00      0.0      0.0
#>                  omu22    omu23    omu24    omu25    omu26    omu27    omu28
#> FullMix_10577 173899.6 318639.9 19363.84 20420.49 58385.43 181823.4 41501.13
#> FullMix_10580 184030.0 320508.9 20265.64 19157.45 48710.46 188771.4 40877.48
#> FullMix_10574 161404.3 412907.4 24725.49 20589.92 48492.77 175498.8 40485.21
#> 50MeOH_10572       0.0      0.0     0.00     0.00     0.00      0.0     0.00
#> 50MeOH_10571       0.0      0.0     0.00     0.00     0.00      0.0     0.00
#> 50MeOH_10570       0.0      0.0     0.00     0.00     0.00      0.0     0.00
#>                 omu29   omu30   omu31   omu32    omu33    omu34    omu35
#> FullMix_10577 4954.48 4900.85 5517.22 5154.74 272380.0 31819.37 12620.44
#> FullMix_10580 5077.47 5479.37 3987.78 5359.76 285909.2 29338.14 12877.17
#> FullMix_10574 5267.41 4986.85 4579.62 5534.43 266564.5 28811.72 12604.05
#> 50MeOH_10572     0.00    0.00    0.00    0.00      0.0     0.00     0.00
#> 50MeOH_10571     0.00    0.00    0.00    0.00      0.0     0.00     0.00
#> 50MeOH_10570     0.00    0.00    0.00    0.00      0.0     0.00     0.00
#>                 omu36    omu37    omu38    omu39    omu40   omu41   omu42
#> FullMix_10577 6277.26 18854.85 46976.79 409198.4 36511.93 6076.86 4468.45
#> FullMix_10580 5916.17 18654.55 52193.91 350554.8 35666.32 6775.01 3774.11
#> FullMix_10574 7034.25 24970.19 47519.93 273212.4 47517.84 5424.17 3847.96
#> 50MeOH_10572     0.00     0.00     0.00      0.0     0.00    0.00    0.00
#> 50MeOH_10571     0.00     0.00     0.00      0.0     0.00    0.00    0.00
#> 50MeOH_10570     0.00     0.00     0.00      0.0     0.00    0.00    0.00
#>                  omu43    omu44    omu45    omu46    omu47    omu48    omu49
#> FullMix_10577 27068.78 652030.4 70964.73 36001.09 161627.7 274910.1 83631.41
#> FullMix_10580 26734.31 683716.2 75555.90 31077.98 152933.7 268262.7 82741.13
#> FullMix_10574 31361.19 651730.6 71424.26 34979.84 176335.1 274001.6 86480.33
#> 50MeOH_10572      0.00      0.0     0.00     0.00      0.0      0.0     0.00
#> 50MeOH_10571      0.00      0.0     0.00     0.00      0.0      0.0     0.00
#> 50MeOH_10570      0.00      0.0     0.00     0.00      0.0      0.0     0.00
#>                 omu50    omu51    omu52    omu53    omu54   omu55    omu56
#> FullMix_10577 4281.21 16711.79 35301.53 12088.39 33270.63 3010.52 26582.23
#> FullMix_10580 4511.35 16938.44 36373.23 12553.59 32493.87 4033.75 27529.64
#> FullMix_10574 4866.85 15092.20 33864.24 11555.30 32586.93 3538.00 26376.34
#> 50MeOH_10572     0.00     0.00     0.00     0.00     0.00    0.00    62.77
#> 50MeOH_10571     0.00     0.00     0.00     0.00     0.00    0.00    66.75
#> 50MeOH_10570     0.00     0.00     0.00     0.00     0.00    0.00    86.03
#>                  omu57   omu58    omu59    omu60    omu61    omu62    omu63
#> FullMix_10577 19217.62 4857.00 27828.94 118849.5 13396.24 33793.49 18244.50
#> FullMix_10580 15021.56 4401.88 30689.97 109969.4 14535.47 29726.62 20318.33
#> FullMix_10574 17376.52 5638.14 33080.43 104191.6 13897.75 27654.14 16688.46
#> 50MeOH_10572      0.00    0.00     0.00      0.0     0.00     0.00     0.00
#> 50MeOH_10571      0.00    0.00     0.00      0.0     0.00     0.00     0.00
#> 50MeOH_10570      0.00    0.00     0.00      0.0     0.00     0.00     0.00
#>                  omu64    omu65    omu66    omu67   omu68   omu69    omu70
#> FullMix_10577 19517.79 18488.68 70084.73 13221.86 4926.15 4032.71 11012.81
#> FullMix_10580 17656.45 17185.37 67041.48 12993.35 6115.98 3372.82  9763.58
#> FullMix_10574 20366.22 20297.28 57873.61 13183.47 4913.68 3227.21 11637.10
#> 50MeOH_10572    745.37     0.00  4252.05     0.00    0.00    0.00     0.00
#> 50MeOH_10571   1056.08     0.00  4058.51     0.00    0.00    0.00     0.00
#> 50MeOH_10570   1022.97     0.00  4472.46     0.00    0.00    0.00     0.00
#>                  omu71    omu72    omu73    omu74    omu75    omu76    omu77
#> FullMix_10577 236016.4 17669.04 21706.97 28646.67 298112.4 17700.28 57529.74
#> FullMix_10580 228756.3 20567.76 20737.98 32781.89 285304.0 18082.23 52614.47
#> FullMix_10574 223554.3 16571.19 20091.00 29263.10 290880.3 20373.03 55657.28
#> 50MeOH_10572       0.0     0.00     0.00     0.00      0.0     0.00     0.00
#> 50MeOH_10571       0.0     0.00     0.00     0.00      0.0     0.00     0.00
#> 50MeOH_10570       0.0     0.00     0.00     0.00      0.0     0.00     0.00
#>                   omu78   omu79   omu80    omu81    omu82   omu83   omu84
#> FullMix_10577  96729.20 8656.43 6420.12 68554.23 385972.3 7590.50 5121.28
#> FullMix_10580  88786.78 8231.63 6524.91 70379.77 381855.8 7732.26 4930.60
#> FullMix_10574 100336.55 9480.51 4643.72 63112.11 401211.1 7826.76 4871.15
#> 50MeOH_10572       0.00    0.00    0.00     0.00      0.0    0.00    0.00
#> 50MeOH_10571       0.00    0.00    0.00     0.00      0.0    0.00    0.00
#> 50MeOH_10570       0.00    0.00    0.00     0.00      0.0    0.00    0.00
#>                 omu85    omu86    omu87     omu88   omu89   omu90   omu91
#> FullMix_10577 7617.49 29334.83 154153.5 1129667.0 5885.18 6365.27 5696.27
#> FullMix_10580 6742.94 27654.73 163650.2  934444.4 6267.79 7226.06 5401.83
#> FullMix_10574 7741.66 24587.89 158731.6  915505.2 6476.30 7316.47 5971.12
#> 50MeOH_10572     0.00     0.00      0.0       0.0    0.00    0.00    0.00
#> 50MeOH_10571     0.00     0.00      0.0       0.0    0.00    0.00    0.00
#> 50MeOH_10570     0.00     0.00      0.0       0.0    0.00    0.00    0.00
#>                  omu92    omu93    omu94    omu95    omu96    omu97    omu98
#> FullMix_10577 69760.48 30055.19 182211.4 108133.2 211819.1 250052.3 178722.7
#> FullMix_10580 82009.32 33768.55 177949.2 116107.8 209578.5 251408.2 176044.7
#> FullMix_10574 81490.64 26904.59 183477.4 114795.2 207711.3 247102.5 175969.0
#> 50MeOH_10572      0.00     0.00      0.0      0.0      0.0      0.0      0.0
#> 50MeOH_10571      0.00     0.00      0.0      0.0      0.0      0.0      0.0
#> 50MeOH_10570      0.00     0.00      0.0      0.0      0.0      0.0      0.0
#>                 omu99   omu100   omu101   omu102    omu103   omu104   omu105
#> FullMix_10577 1601.46  7858.40 124858.4 118422.5 118213.27 14911.96 11891.37
#> FullMix_10580 1822.29 10044.64 113163.8 114335.3 118710.73 16099.34 11496.39
#> FullMix_10574 1935.13  7032.65 140301.4 134025.3  83679.05 16647.38 13714.92
#> 50MeOH_10572     0.00     0.00      0.0      0.0      0.00     0.00     0.00
#> 50MeOH_10571     0.00     0.00      0.0      0.0      0.00     0.00     0.00
#> 50MeOH_10570     0.00     0.00      0.0      0.0      0.00     0.00     0.00
#>                 omu106   omu107   omu108   omu109    omu110   omu111   omu112
#> FullMix_10577 16394.66 41318.40 11251.71 189467.1 186699.50 78223.05 43702.04
#> FullMix_10580 16997.37 38294.09 10179.66 181452.4 167348.86 84049.83 42647.61
#> FullMix_10574 17020.84 50491.21  8777.77 192459.9 206364.25 79450.47 47851.07
#> 50MeOH_10572      0.00     0.00     0.00      0.0   4854.20     0.00     0.00
#> 50MeOH_10571      0.00     0.00     0.00      0.0   4079.94     0.00     0.00
#> 50MeOH_10570      0.00     0.00     0.00      0.0   6007.28     0.00     0.00
#>                 omu113   omu114   omu115   omu116   omu117    omu118  omu119
#> FullMix_10577 342213.5 47293.55 13856.37 39554.72 11122.01 162278.70 2388932
#> FullMix_10580 349856.8 49961.25 13669.61 46702.82 13464.26 152789.64 2212812
#> FullMix_10574 382052.3 61257.75 16411.03 36884.52 10273.41 110851.04 1938057
#> 50MeOH_10572       0.0     0.00     0.00     0.00     0.00   1823.84       0
#> 50MeOH_10571       0.0     0.00     0.00     0.00     0.00   1630.82       0
#> 50MeOH_10570       0.0     0.00     0.00     0.00     0.00   1333.93       0
#>                 omu120   omu121   omu122   omu123   omu124   omu125   omu126
#> FullMix_10577 680710.4 246523.6 450185.8 45833.30 31928.30 23024.84 30193161
#> FullMix_10580 688871.6 245164.4 400433.1 45543.40 31017.15 19896.01 30146821
#> FullMix_10574 947038.7 215962.4 412101.9 47552.48 23979.63 17154.91 29499959
#> 50MeOH_10572       0.0      0.0      0.0     0.00     0.00     0.00        0
#> 50MeOH_10571       0.0      0.0      0.0     0.00     0.00     0.00        0
#> 50MeOH_10570       0.0      0.0      0.0     0.00     0.00     0.00        0
#>                 omu127     omu128   omu129  omu130  omu131    omu132    omu133
#> FullMix_10577 10769016 9786368.01 11710055 1149551 4643096 156986.76 103601.67
#> FullMix_10580 10775346 9832396.75 11715098 1173830 4805970 150261.84  99383.66
#> FullMix_10574 10807491 8798990.64 11352230 1434818 4996139 153891.42 103353.06
#> 50MeOH_10572         0     120.02        0       0       0   4735.21      0.00
#> 50MeOH_10571         0     137.23        0       0       0   4497.09      0.00
#> 50MeOH_10570         0     131.43        0       0       0   5735.87      0.00
#>                 omu134   omu135   omu136   omu137   omu138   omu139   omu140
#> FullMix_10577 557606.8 238482.5 303676.8 112441.9 266260.0 128456.0 328626.3
#> FullMix_10580 563725.1 226279.8 330701.2 110115.1 295186.3 105812.7 294291.8
#> FullMix_10574 555530.1 253784.6 347816.7 119541.2 295237.4 126295.4 336995.4
#> 50MeOH_10572       0.0      0.0      0.0      0.0      0.0      0.0      0.0
#> 50MeOH_10571       0.0      0.0      0.0      0.0      0.0      0.0      0.0
#> 50MeOH_10570       0.0      0.0      0.0      0.0      0.0      0.0      0.0
#>                 omu141  omu142   omu143  omu144   omu145   omu146   omu147
#> FullMix_10577 554515.4 4380434 58547.07 6967279 24579.75 330729.5 482430.5
#> FullMix_10580 561919.6 3743478 60980.05 6329251 27237.91 326762.2 445387.5
#> FullMix_10574 546595.5 3941644 59644.13 6617553 24137.47 239295.9 414433.3
#> 50MeOH_10572       0.0       0     0.00       0     0.00      0.0      0.0
#> 50MeOH_10571       0.0       0     0.00       0     0.00      0.0      0.0
#> 50MeOH_10570       0.0       0     0.00       0     0.00      0.0      0.0
#>                 omu148   omu149   omu150   omu151  omu152     omu153
#> FullMix_10577 23360.14 187580.6 455963.4 37289.67 4458107 3025208.45
#> FullMix_10580 17366.65 165105.4 541673.6 31631.85 4239598 3215001.06
#> FullMix_10574 25053.21 159111.4 465740.5 35837.54 4202890 3298316.07
#> 50MeOH_10572      0.00      0.0      0.0     0.00       0     241.30
#> 50MeOH_10571      0.00      0.0      0.0     0.00       0     201.07
#> 50MeOH_10570      0.00      0.0      0.0     0.00       0     238.39