Rarefy MS1 Feature Table

rarefy_ms() performs a single subsampling of MS1 features in sample. Feature intensities are subsampled to the supplied size and accounts for intensity thresholds due to machine limits and background noise. Specifically, features whose abundance falls below the threshold after rarefying are removed. This allows for accurate representation of samples at different dilutions regardless of the desired submsampling size.

Usage

rarefy_ms(
  community_object,
  size,
  threshold,
  number_of_threads = detectCores(),
  seed = 123
)

Arguments

community_object

A community_object

size

The desired total sample intensity to subsample to.

threshold

The individual feature threshold. Each subsampled feature must be >= this value to be retained.

number_of_threads

the number of threads you wish to use for this calculation. Defaults to the number of threads on your computer.

seed

the RNG (random number generator) seed you would like to use.

Value

A external_pointer that references a community matrix of rarefied feature intensities.

returns a matrix object that contains your rarefied data.

Examples

data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")

matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)
#> [1] "Reading: /home/runner/work/_temp/Library/mums2/extdata/botryllus_v2.gnps.mgf ..."
#> Computing                                                    | 0%  ETA: -...Computing ■                                                  | 2%  ETA: ...Computing ■■                                                 | 4%  ETA: ...Computing ■■■                                                | 6%  ETA: ...Computing ■■■■                                               | 8%  ETA: ...Computing ■■■■■                                              | 10%  ETA: ...Computing ■■■■■■                                             | 12%  ETA: ...Computing ■■■■■■■                                            | 14%  ETA: ...Computing ■■■■■■■■                                           | 16%  ETA: ...Computing ■■■■■■■■■                                          | 18%  ETA: ...Computing ■■■■■■■■■■                                         | 20%  ETA: ...Computing ■■■■■■■■■■■                                        | 22%  ETA: ...Computing ■■■■■■■■■■■■                                       | 24%  ETA: ...Computing ■■■■■■■■■■■■■                                      | 26%  ETA: ...Computing ■■■■■■■■■■■■■■                                     | 28%  ETA: ...Computing ■■■■■■■■■■■■■■■                                    | 30%  ETA: ...Computing ■■■■■■■■■■■■■■■■                                   | 32%  ETA: ...Computing ■■■■■■■■■■■■■■■■■                                  | 34%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■                                 | 36%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■                                | 38%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■                               | 40%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■                              | 42%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■                             | 44%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■                            | 46%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■                           | 48%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■                          | 50%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■                         | 52%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■                        | 54%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■                       | 56%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                      | 58%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                     | 60%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                    | 62%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                   | 64%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                  | 66%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                 | 68%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                | 70%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■               | 72%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■              | 74%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■             | 76%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            | 78%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■           | 80%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■          | 82%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■         | 84%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■        | 86%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■       | 88%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■      | 90%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     | 92%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    | 94%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   | 96%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  | 98%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ | 100%  ETA: ...
#> [1] "17/349 peaks have an MS2 spectra."

community_object <- create_community_matrix_object(matched_data)
rarefy_ms(community_object, 400, 10)
#>                            559.37264 Da 448.87 s 882.45822 Da 458.82 s
#> 221012_DGM_Blank1_1_1_390                     17                    14
#> 221012_DGM_Blank1_1_2_391                      0                     0
#> 221012_DGM_Blank1_1_3_392                      0                     0
#> 221012_DGM_MB1588_3_1_395                      0                    16
#> 221012_DGM_MB1588_3_2_396                      0                     0
#> 221012_DGM_MB1588_3_3_397                      0                    27
#> 221012_DGM_MB1589_4_1_398                      0                    25
#> 221012_DGM_MB1589_4_2_399                      0                    11
#> 221012_DGM_MB1589_4_3_400                      0                     0
#> 221012_DGM_MB1590_5_1_401                      0                    10
#> 221012_DGM_MB1590_5_2_402                      0                     0
#> 221012_DGM_MB1590_5_3_403                      0                    33
#> 221012_DGM_Blank2_1_1_404                      0                     0
#> 221012_DGM_Blank2_1_2_405                      0                     0
#> 221012_DGM_Blank2_1_3_406                      0                    17
#> 221012_DGM_MB1591_6_1_407                     26                    29
#> 221012_DGM_MB1591_6_2_408                      0                    16
#> 221012_DGM_MB1591_6_3_409                      0                    48
#> 221012_DGM_MB1592_7_1_410                     29                    63
#> 221012_DGM_MB1592_7_2_411                     20                    23
#> 221012_DGM_MB1592_7_3_412                     13                    29
#> 221012_DGM_MB1593_8_1_413                     12                    28
#> 221012_DGM_MB1593_8_2_414                      0                    29
#> 221012_DGM_MB1593_8_3_415                      0                    17
#> 221012_DGM_MB1594_9_1_416                     32                     0
#> 221012_DGM_MB1594_9_2_417                      0                    15
#> 221012_DGM_MB1594_9_3_418                      0                     0
#> 221012_DGM_Blank3_1_1_419                      0                    21
#> 221012_DGM_Blank3_1_2_420                      0                    67
#> 221012_DGM_Blank3_1_3_421                      0                    56
#> 221012_DGM_MB1595_10_1_422                    24                    32
#> 221012_DGM_MB1595_10_2_423                    58                    47
#> 221012_DGM_MB1595_10_3_424                     0                    37
#> 221012_DGM_MB1597_11_1_425                     0                   206
#> 221012_DGM_MB1597_11_2_426                    20                   205
#> 221012_DGM_MB1597_11_3_427                     0                   304
#> 221012_DGM_MB1598_12_1_428                     0                   117
#> 221012_DGM_MB1598_12_2_429                     0                   173
#> 221012_DGM_MB1598_12_3_430                     0                   155
#> 221012_DGM_MB1599_13_1_431                    12                    74
#> 221012_DGM_MB1599_13_2_432                    16                    96
#> 221012_DGM_MB1599_13_3_433                    17                   111
#> 221012_DGM_Blank4_1_1_434                     27                    18
#> 221012_DGM_Blank4_1_2_435                      0                    27
#> 221012_DGM_Blank4_1_3_436                      0                    21
#>                            782.56450 Da 436.43 s 536.52815 Da 486.44 s
#> 221012_DGM_Blank1_1_1_390                      0                    64
#> 221012_DGM_Blank1_1_2_391                      0                    10
#> 221012_DGM_Blank1_1_3_392                      0                    18
#> 221012_DGM_MB1588_3_1_395                      0                    56
#> 221012_DGM_MB1588_3_2_396                      0                    22
#> 221012_DGM_MB1588_3_3_397                      0                     0
#> 221012_DGM_MB1589_4_1_398                      0                    53
#> 221012_DGM_MB1589_4_2_399                     27                    47
#> 221012_DGM_MB1589_4_3_400                      0                    88
#> 221012_DGM_MB1590_5_1_401                      0                    32
#> 221012_DGM_MB1590_5_2_402                      0                     0
#> 221012_DGM_MB1590_5_3_403                      0                    95
#> 221012_DGM_Blank2_1_1_404                      0                     0
#> 221012_DGM_Blank2_1_2_405                      0                    46
#> 221012_DGM_Blank2_1_3_406                      0                    87
#> 221012_DGM_MB1591_6_1_407                      0                    54
#> 221012_DGM_MB1591_6_2_408                    142                    58
#> 221012_DGM_MB1591_6_3_409                    157                   124
#> 221012_DGM_MB1592_7_1_410                      0                     0
#> 221012_DGM_MB1592_7_2_411                      0                     0
#> 221012_DGM_MB1592_7_3_412                      0                     0
#> 221012_DGM_MB1593_8_1_413                      0                     0
#> 221012_DGM_MB1593_8_2_414                      0                    29
#> 221012_DGM_MB1593_8_3_415                     48                    65
#> 221012_DGM_MB1594_9_1_416                     74                    66
#> 221012_DGM_MB1594_9_2_417                      0                    65
#> 221012_DGM_MB1594_9_3_418                     74                     0
#> 221012_DGM_Blank3_1_1_419                      0                     0
#> 221012_DGM_Blank3_1_2_420                    180                     0
#> 221012_DGM_Blank3_1_3_421                      0                     0
#> 221012_DGM_MB1595_10_1_422                     0                     0
#> 221012_DGM_MB1595_10_2_423                     0                     0
#> 221012_DGM_MB1595_10_3_424                     0                     0
#> 221012_DGM_MB1597_11_1_425                    53                     0
#> 221012_DGM_MB1597_11_2_426                    53                    31
#> 221012_DGM_MB1597_11_3_427                    32                     0
#> 221012_DGM_MB1598_12_1_428                    90                     0
#> 221012_DGM_MB1598_12_2_429                    84                     0
#> 221012_DGM_MB1598_12_3_430                   108                     0
#> 221012_DGM_MB1599_13_1_431                    56                    29
#> 221012_DGM_MB1599_13_2_432                    55                    47
#> 221012_DGM_MB1599_13_3_433                     0                   112
#> 221012_DGM_Blank4_1_1_434                      0                     0
#> 221012_DGM_Blank4_1_2_435                      0                   147
#> 221012_DGM_Blank4_1_3_436                     81                     0
#>                            648.44226 Da 504.36 s 609.17531 Da 532.76 s
#> 221012_DGM_Blank1_1_1_390                      0                     0
#> 221012_DGM_Blank1_1_2_391                      0                   360
#> 221012_DGM_Blank1_1_3_392                      0                   265
#> 221012_DGM_MB1588_3_1_395                      0                     0
#> 221012_DGM_MB1588_3_2_396                     42                   226
#> 221012_DGM_MB1588_3_3_397                      0                     0
#> 221012_DGM_MB1589_4_1_398                     58                     0
#> 221012_DGM_MB1589_4_2_399                     45                     0
#> 221012_DGM_MB1589_4_3_400                    106                     0
#> 221012_DGM_MB1590_5_1_401                      0                   201
#> 221012_DGM_MB1590_5_2_402                     10                   375
#> 221012_DGM_MB1590_5_3_403                      0                     0
#> 221012_DGM_Blank2_1_1_404                      0                     0
#> 221012_DGM_Blank2_1_2_405                      0                   295
#> 221012_DGM_Blank2_1_3_406                      0                     0
#> 221012_DGM_MB1591_6_1_407                    151                     0
#> 221012_DGM_MB1591_6_2_408                      0                   122
#> 221012_DGM_MB1591_6_3_409                      0                     0
#> 221012_DGM_MB1592_7_1_410                      0                     0
#> 221012_DGM_MB1592_7_2_411                      0                     0
#> 221012_DGM_MB1592_7_3_412                      0                    16
#> 221012_DGM_MB1593_8_1_413                     55                    59
#> 221012_DGM_MB1593_8_2_414                     93                    56
#> 221012_DGM_MB1593_8_3_415                      0                   104
#> 221012_DGM_MB1594_9_1_416                     68                     0
#> 221012_DGM_MB1594_9_2_417                      0                     0
#> 221012_DGM_MB1594_9_3_418                      0                    24
#> 221012_DGM_Blank3_1_1_419                     86                    73
#> 221012_DGM_Blank3_1_2_420                      0                     0
#> 221012_DGM_Blank3_1_3_421                      0                   163
#> 221012_DGM_MB1595_10_1_422                     0                    55
#> 221012_DGM_MB1595_10_2_423                     0                     0
#> 221012_DGM_MB1595_10_3_424                     0                     0
#> 221012_DGM_MB1597_11_1_425                    11                    10
#> 221012_DGM_MB1597_11_2_426                     0                     0
#> 221012_DGM_MB1597_11_3_427                     0                     0
#> 221012_DGM_MB1598_12_1_428                     0                    21
#> 221012_DGM_MB1598_12_2_429                     0                     0
#> 221012_DGM_MB1598_12_3_430                     0                     0
#> 221012_DGM_MB1599_13_1_431                    64                    27
#> 221012_DGM_MB1599_13_2_432                     0                    70
#> 221012_DGM_MB1599_13_3_433                     0                     0
#> 221012_DGM_Blank4_1_1_434                      0                   219
#> 221012_DGM_Blank4_1_2_435                      0                   152
#> 221012_DGM_Blank4_1_3_436                      0                    74
#>                            647.41630 Da 447.95 s 778.53241 Da 445.03 s
#> 221012_DGM_Blank1_1_1_390                      0                     0
#> 221012_DGM_Blank1_1_2_391                      0                     0
#> 221012_DGM_Blank1_1_3_392                      0                     0
#> 221012_DGM_MB1588_3_1_395                     32                    53
#> 221012_DGM_MB1588_3_2_396                      0                     0
#> 221012_DGM_MB1588_3_3_397                      0                     0
#> 221012_DGM_MB1589_4_1_398                     14                     0
#> 221012_DGM_MB1589_4_2_399                     10                     0
#> 221012_DGM_MB1589_4_3_400                      0                     0
#> 221012_DGM_MB1590_5_1_401                      0                    73
#> 221012_DGM_MB1590_5_2_402                      0                     0
#> 221012_DGM_MB1590_5_3_403                      0                     0
#> 221012_DGM_Blank2_1_1_404                      0                   249
#> 221012_DGM_Blank2_1_2_405                      0                    59
#> 221012_DGM_Blank2_1_3_406                     39                   174
#> 221012_DGM_MB1591_6_1_407                      0                     0
#> 221012_DGM_MB1591_6_2_408                      0                     0
#> 221012_DGM_MB1591_6_3_409                      0                     0
#> 221012_DGM_MB1592_7_1_410                      0                     0
#> 221012_DGM_MB1592_7_2_411                      0                     0
#> 221012_DGM_MB1592_7_3_412                      0                     0
#> 221012_DGM_MB1593_8_1_413                      0                     0
#> 221012_DGM_MB1593_8_2_414                      0                     0
#> 221012_DGM_MB1593_8_3_415                      0                     0
#> 221012_DGM_MB1594_9_1_416                      0                     0
#> 221012_DGM_MB1594_9_2_417                     21                    78
#> 221012_DGM_MB1594_9_3_418                      0                   167
#> 221012_DGM_Blank3_1_1_419                     21                     0
#> 221012_DGM_Blank3_1_2_420                      0                     0
#> 221012_DGM_Blank3_1_3_421                      0                     0
#> 221012_DGM_MB1595_10_1_422                     0                     0
#> 221012_DGM_MB1595_10_2_423                     0                     0
#> 221012_DGM_MB1595_10_3_424                     0                     0
#> 221012_DGM_MB1597_11_1_425                    19                    16
#> 221012_DGM_MB1597_11_2_426                    27                    17
#> 221012_DGM_MB1597_11_3_427                    25                     0
#> 221012_DGM_MB1598_12_1_428                     0                    57
#> 221012_DGM_MB1598_12_2_429                    36                     0
#> 221012_DGM_MB1598_12_3_430                    47                     0
#> 221012_DGM_MB1599_13_1_431                     0                    22
#> 221012_DGM_MB1599_13_2_432                     0                     0
#> 221012_DGM_MB1599_13_3_433                    53                    36
#> 221012_DGM_Blank4_1_1_434                      0                     0
#> 221012_DGM_Blank4_1_2_435                      0                     0
#> 221012_DGM_Blank4_1_3_436                      0                    39
#>                            555.39892 Da 347.32 s 761.54191 Da 605.01 s
#> 221012_DGM_Blank1_1_1_390                     14                    84
#> 221012_DGM_Blank1_1_2_391                      0                     0
#> 221012_DGM_Blank1_1_3_392                      0                     0
#> 221012_DGM_MB1588_3_1_395                     12                   213
#> 221012_DGM_MB1588_3_2_396                      0                    95
#> 221012_DGM_MB1588_3_3_397                      0                   282
#> 221012_DGM_MB1589_4_1_398                      0                   235
#> 221012_DGM_MB1589_4_2_399                      0                   167
#> 221012_DGM_MB1589_4_3_400                     16                   138
#> 221012_DGM_MB1590_5_1_401                      0                    56
#> 221012_DGM_MB1590_5_2_402                      0                    15
#> 221012_DGM_MB1590_5_3_403                     10                   148
#> 221012_DGM_Blank2_1_1_404                     27                    25
#> 221012_DGM_Blank2_1_2_405                      0                     0
#> 221012_DGM_Blank2_1_3_406                     16                    67
#> 221012_DGM_MB1591_6_1_407                      0                   104
#> 221012_DGM_MB1591_6_2_408                      0                    62
#> 221012_DGM_MB1591_6_3_409                     26                     0
#> 221012_DGM_MB1592_7_1_410                     17                   115
#> 221012_DGM_MB1592_7_2_411                      0                    73
#> 221012_DGM_MB1592_7_3_412                     12                   104
#> 221012_DGM_MB1593_8_1_413                      0                    64
#> 221012_DGM_MB1593_8_2_414                     29                    99
#> 221012_DGM_MB1593_8_3_415                     15                    84
#> 221012_DGM_MB1594_9_1_416                     25                     0
#> 221012_DGM_MB1594_9_2_417                     11                    29
#> 221012_DGM_MB1594_9_3_418                     21                    26
#> 221012_DGM_Blank3_1_1_419                     20                    43
#> 221012_DGM_Blank3_1_2_420                     45                    88
#> 221012_DGM_Blank3_1_3_421                     69                   102
#> 221012_DGM_MB1595_10_1_422                    13                   134
#> 221012_DGM_MB1595_10_2_423                    15                   124
#> 221012_DGM_MB1595_10_3_424                     0                   157
#> 221012_DGM_MB1597_11_1_425                     0                     0
#> 221012_DGM_MB1597_11_2_426                    16                    12
#> 221012_DGM_MB1597_11_3_427                    17                     0
#> 221012_DGM_MB1598_12_1_428                     0                    27
#> 221012_DGM_MB1598_12_2_429                    27                    24
#> 221012_DGM_MB1598_12_3_430                    25                    51
#> 221012_DGM_MB1599_13_1_431                     0                    45
#> 221012_DGM_MB1599_13_2_432                    18                    65
#> 221012_DGM_MB1599_13_3_433                    35                    21
#> 221012_DGM_Blank4_1_1_434                     11                    33
#> 221012_DGM_Blank4_1_2_435                     18                    28
#> 221012_DGM_Blank4_1_3_436                     12                     0
#>                            687.48123 Da 402.80 s 1298.31327 Da 555.75 s
#> 221012_DGM_Blank1_1_1_390                      0                     64
#> 221012_DGM_Blank1_1_2_391                      0                     20
#> 221012_DGM_Blank1_1_3_392                      0                     42
#> 221012_DGM_MB1588_3_1_395                     18                      0
#> 221012_DGM_MB1588_3_2_396                      0                      0
#> 221012_DGM_MB1588_3_3_397                     30                      0
#> 221012_DGM_MB1589_4_1_398                      0                      0
#> 221012_DGM_MB1589_4_2_399                      0                      0
#> 221012_DGM_MB1589_4_3_400                     23                      0
#> 221012_DGM_MB1590_5_1_401                     15                      0
#> 221012_DGM_MB1590_5_2_402                      0                      0
#> 221012_DGM_MB1590_5_3_403                     56                      0
#> 221012_DGM_Blank2_1_1_404                      0                      0
#> 221012_DGM_Blank2_1_2_405                      0                      0
#> 221012_DGM_Blank2_1_3_406                      0                      0
#> 221012_DGM_MB1591_6_1_407                      0                      0
#> 221012_DGM_MB1591_6_2_408                      0                      0
#> 221012_DGM_MB1591_6_3_409                      0                      0
#> 221012_DGM_MB1592_7_1_410                    159                      0
#> 221012_DGM_MB1592_7_2_411                    147                      0
#> 221012_DGM_MB1592_7_3_412                    152                      0
#> 221012_DGM_MB1593_8_1_413                     57                      0
#> 221012_DGM_MB1593_8_2_414                     65                      0
#> 221012_DGM_MB1593_8_3_415                     49                      0
#> 221012_DGM_MB1594_9_1_416                    116                      0
#> 221012_DGM_MB1594_9_2_417                     65                      0
#> 221012_DGM_MB1594_9_3_418                     88                      0
#> 221012_DGM_Blank3_1_1_419                      0                      0
#> 221012_DGM_Blank3_1_2_420                      0                      0
#> 221012_DGM_Blank3_1_3_421                      0                      0
#> 221012_DGM_MB1595_10_1_422                    85                      0
#> 221012_DGM_MB1595_10_2_423                   104                      0
#> 221012_DGM_MB1595_10_3_424                    93                      0
#> 221012_DGM_MB1597_11_1_425                    24                      0
#> 221012_DGM_MB1597_11_2_426                    19                      0
#> 221012_DGM_MB1597_11_3_427                    22                      0
#> 221012_DGM_MB1598_12_1_428                    51                      0
#> 221012_DGM_MB1598_12_2_429                    56                      0
#> 221012_DGM_MB1598_12_3_430                     0                      0
#> 221012_DGM_MB1599_13_1_431                    60                      0
#> 221012_DGM_MB1599_13_2_432                     0                      0
#> 221012_DGM_MB1599_13_3_433                     0                      0
#> 221012_DGM_Blank4_1_1_434                      0                      0
#> 221012_DGM_Blank4_1_2_435                      0                      0
#> 221012_DGM_Blank4_1_3_436                    131                      0
#>                            794.53451 Da 530.10 s 1293.35540 Da 486.71 s
#> 221012_DGM_Blank1_1_1_390                      0                    117
#> 221012_DGM_Blank1_1_2_391                      0                     17
#> 221012_DGM_Blank1_1_3_392                      0                     15
#> 221012_DGM_MB1588_3_1_395                      0                      0
#> 221012_DGM_MB1588_3_2_396                     15                      0
#> 221012_DGM_MB1588_3_3_397                     61                      0
#> 221012_DGM_MB1589_4_1_398                      0                      0
#> 221012_DGM_MB1589_4_2_399                      0                      0
#> 221012_DGM_MB1589_4_3_400                      0                      0
#> 221012_DGM_MB1590_5_1_401                      0                      0
#> 221012_DGM_MB1590_5_2_402                      0                      0
#> 221012_DGM_MB1590_5_3_403                      0                     28
#> 221012_DGM_Blank2_1_1_404                     99                      0
#> 221012_DGM_Blank2_1_2_405                      0                      0
#> 221012_DGM_Blank2_1_3_406                      0                      0
#> 221012_DGM_MB1591_6_1_407                     36                      0
#> 221012_DGM_MB1591_6_2_408                      0                      0
#> 221012_DGM_MB1591_6_3_409                     18                      0
#> 221012_DGM_MB1592_7_1_410                      0                      0
#> 221012_DGM_MB1592_7_2_411                      0                      0
#> 221012_DGM_MB1592_7_3_412                      0                      0
#> 221012_DGM_MB1593_8_1_413                      0                      0
#> 221012_DGM_MB1593_8_2_414                      0                      0
#> 221012_DGM_MB1593_8_3_415                      0                      0
#> 221012_DGM_MB1594_9_1_416                      0                      0
#> 221012_DGM_MB1594_9_2_417                      0                      0
#> 221012_DGM_MB1594_9_3_418                      0                      0
#> 221012_DGM_Blank3_1_1_419                      0                      0
#> 221012_DGM_Blank3_1_2_420                      0                      0
#> 221012_DGM_Blank3_1_3_421                      0                      0
#> 221012_DGM_MB1595_10_1_422                     0                      0
#> 221012_DGM_MB1595_10_2_423                     0                      0
#> 221012_DGM_MB1595_10_3_424                     0                      0
#> 221012_DGM_MB1597_11_1_425                     0                      0
#> 221012_DGM_MB1597_11_2_426                     0                      0
#> 221012_DGM_MB1597_11_3_427                     0                      0
#> 221012_DGM_MB1598_12_1_428                     0                      0
#> 221012_DGM_MB1598_12_2_429                     0                      0
#> 221012_DGM_MB1598_12_3_430                    14                      0
#> 221012_DGM_MB1599_13_1_431                     0                      0
#> 221012_DGM_MB1599_13_2_432                    18                      0
#> 221012_DGM_MB1599_13_3_433                    10                      0
#> 221012_DGM_Blank4_1_1_434                     17                      0
#> 221012_DGM_Blank4_1_2_435                     18                      0
#> 221012_DGM_Blank4_1_3_436                     11                      0
#>                            677.46670 Da 408.06 s 687.48111 Da 390.53 s
#> 221012_DGM_Blank1_1_1_390                     26                     0
#> 221012_DGM_Blank1_1_2_391                      0                     0
#> 221012_DGM_Blank1_1_3_392                      0                     0
#> 221012_DGM_MB1588_3_1_395                      0                     0
#> 221012_DGM_MB1588_3_2_396                      0                     0
#> 221012_DGM_MB1588_3_3_397                      0                     0
#> 221012_DGM_MB1589_4_1_398                     15                     0
#> 221012_DGM_MB1589_4_2_399                      0                     0
#> 221012_DGM_MB1589_4_3_400                      0                     0
#> 221012_DGM_MB1590_5_1_401                     13                     0
#> 221012_DGM_MB1590_5_2_402                      0                     0
#> 221012_DGM_MB1590_5_3_403                     30                     0
#> 221012_DGM_Blank2_1_1_404                      0                     0
#> 221012_DGM_Blank2_1_2_405                      0                     0
#> 221012_DGM_Blank2_1_3_406                      0                     0
#> 221012_DGM_MB1591_6_1_407                      0                     0
#> 221012_DGM_MB1591_6_2_408                      0                     0
#> 221012_DGM_MB1591_6_3_409                      0                     0
#> 221012_DGM_MB1592_7_1_410                      0                     0
#> 221012_DGM_MB1592_7_2_411                      0                   137
#> 221012_DGM_MB1592_7_3_412                      0                    74
#> 221012_DGM_MB1593_8_1_413                     59                    44
#> 221012_DGM_MB1593_8_2_414                      0                     0
#> 221012_DGM_MB1593_8_3_415                      0                     0
#> 221012_DGM_MB1594_9_1_416                      0                     0
#> 221012_DGM_MB1594_9_2_417                     40                    49
#> 221012_DGM_MB1594_9_3_418                      0                     0
#> 221012_DGM_Blank3_1_1_419                    119                     0
#> 221012_DGM_Blank3_1_2_420                      0                     0
#> 221012_DGM_Blank3_1_3_421                      0                     0
#> 221012_DGM_MB1595_10_1_422                     0                    57
#> 221012_DGM_MB1595_10_2_423                     0                    52
#> 221012_DGM_MB1595_10_3_424                    76                    37
#> 221012_DGM_MB1597_11_1_425                    53                    16
#> 221012_DGM_MB1597_11_2_426                     0                     0
#> 221012_DGM_MB1597_11_3_427                     0                     0
#> 221012_DGM_MB1598_12_1_428                     0                    37
#> 221012_DGM_MB1598_12_2_429                     0                     0
#> 221012_DGM_MB1598_12_3_430                     0                     0
#> 221012_DGM_MB1599_13_1_431                     0                     0
#> 221012_DGM_MB1599_13_2_432                     0                     0
#> 221012_DGM_MB1599_13_3_433                     0                     0
#> 221012_DGM_Blank4_1_1_434                     61                     0
#> 221012_DGM_Blank4_1_2_435                      0                     0
#> 221012_DGM_Blank4_1_3_436                     18                     0
#>                            872.25849 Da 530.21 s
#> 221012_DGM_Blank1_1_1_390                      0
#> 221012_DGM_Blank1_1_2_391                      0
#> 221012_DGM_Blank1_1_3_392                     60
#> 221012_DGM_MB1588_3_1_395                      0
#> 221012_DGM_MB1588_3_2_396                      0
#> 221012_DGM_MB1588_3_3_397                      0
#> 221012_DGM_MB1589_4_1_398                      0
#> 221012_DGM_MB1589_4_2_399                    102
#> 221012_DGM_MB1589_4_3_400                     29
#> 221012_DGM_MB1590_5_1_401                      0
#> 221012_DGM_MB1590_5_2_402                      0
#> 221012_DGM_MB1590_5_3_403                      0
#> 221012_DGM_Blank2_1_1_404                      0
#> 221012_DGM_Blank2_1_2_405                      0
#> 221012_DGM_Blank2_1_3_406                      0
#> 221012_DGM_MB1591_6_1_407                      0
#> 221012_DGM_MB1591_6_2_408                      0
#> 221012_DGM_MB1591_6_3_409                     27
#> 221012_DGM_MB1592_7_1_410                     17
#> 221012_DGM_MB1592_7_2_411                      0
#> 221012_DGM_MB1592_7_3_412                      0
#> 221012_DGM_MB1593_8_1_413                     22
#> 221012_DGM_MB1593_8_2_414                      0
#> 221012_DGM_MB1593_8_3_415                     18
#> 221012_DGM_MB1594_9_1_416                     19
#> 221012_DGM_MB1594_9_2_417                     27
#> 221012_DGM_MB1594_9_3_418                      0
#> 221012_DGM_Blank3_1_1_419                     17
#> 221012_DGM_Blank3_1_2_420                     20
#> 221012_DGM_Blank3_1_3_421                     10
#> 221012_DGM_MB1595_10_1_422                     0
#> 221012_DGM_MB1595_10_2_423                     0
#> 221012_DGM_MB1595_10_3_424                     0
#> 221012_DGM_MB1597_11_1_425                     0
#> 221012_DGM_MB1597_11_2_426                     0
#> 221012_DGM_MB1597_11_3_427                     0
#> 221012_DGM_MB1598_12_1_428                     0
#> 221012_DGM_MB1598_12_2_429                     0
#> 221012_DGM_MB1598_12_3_430                     0
#> 221012_DGM_MB1599_13_1_431                    11
#> 221012_DGM_MB1599_13_2_432                    15
#> 221012_DGM_MB1599_13_3_433                    13
#> 221012_DGM_Blank4_1_1_434                     14
#> 221012_DGM_Blank4_1_2_435                     10
#> 221012_DGM_Blank4_1_3_436                     13