Print Community Object

S3 function for print the community object

Usage

# S3 method for class 'community_object'
print(x, ...)

Arguments

x

the object created from the create_community_object() function.

...

other parameters that are included in the print function.

Examples

data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   meta_data =
                   mums2::mums2_example("meta_data_boryillus.csv"),
                   format = "None")
#> If peak table has corrupted compound names they will be converted to
#>       utf-8 and if there are any commas, they will be converted to periods(.).

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())
#> ℹ Checking 1500 peaks for mispicked peaks.
#> ℹ Argument merge_peaks is: TRUE. Merging mispicked peaks with method sum.
#> ✔ 50 ions failed the mispicked filter, 1450 ions remain.
#> ℹ Parsing 1450 peaks for replicability across technical replicates.
#> ✔ 329 ions failed the cv_filter filter, 1121 ions remain.
#> ℹ Parsing 1121 peaks based on the sample group: Blanks.
#> ℹ Argument remove_ions is: TRUE.Removing peaks from Blanks.
#> ✔ 342 ions failed the Blanks filter, 779 ions remain.
#> ℹ Parsing 779 peaks for insource ions.
#> ✔ 55 ions failed the insource filter, 724 ions remain.

change_rt_to_seconds_or_minute(filtered_data, "minutes")
#> [1] "Changing rt values to minutes"
#> Key: <Compound, mz, kmd, RTINMINUTES>
#>                    Compound        mz     kmd RTINMINUTES
#>                      <char>     <num>   <num>       <num>
#>   1: 1000.65345 Da 418.99 s 1001.6607 0.66073   0.1163333
#>   2: 1002.23833 Da 440.75 s 1003.2456 0.24560   0.1225000
#>   3: 1002.57692 Da 414.74 s 1003.5842 0.58419   0.1151667
#>   4: 1004.19672 Da 603.96 s 1005.2040 0.20400   0.1678333
#>   5: 1004.23580 Da 548.28 s 1005.2431 0.24307   0.1523333
#>  ---                                                     
#> 720:  980.84769 Da 515.58 s  981.8550 0.85497   0.1431667
#> 721:  984.23693 Da 604.23 s  985.2442 0.24420   0.1678333
#> 722:  984.65546 Da 460.05 s  985.6627 0.66274   0.1278333
#> 723:  997.30596 Da 605.90 s  998.3132 0.31323   0.1683333
#> 724:  998.28767 Da 556.84 s  981.2843 0.28429   0.1546667
#>      221012_DGM_Blank1_1_1_390 221012_DGM_Blank1_1_2_391
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank1_1_3_392 221012_DGM_Blank2_1_1_404
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank2_1_2_405 221012_DGM_Blank2_1_3_406
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank3_1_1_419 221012_DGM_Blank3_1_2_420
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank3_1_3_421 221012_DGM_Blank4_1_1_434
#>                          <num>                     <num>
#>   1:                         0                   1538.23
#>   2:                         0                      0.00
#>   3:                         0                      0.00
#>   4:                         0                      0.00
#>   5:                         0                      0.00
#>  ---                                                    
#> 720:                         0                      0.00
#> 721:                         0                      0.00
#> 722:                         0                      0.00
#> 723:                         0                      0.00
#> 724:                         0                      0.00
#>      221012_DGM_Blank4_1_2_435 221012_DGM_Blank4_1_3_436
#>                          <num>                     <num>
#>   1:                  1201.261                  1180.144
#>   2:                     0.000                     0.000
#>   3:                     0.000                     0.000
#>   4:                     0.000                     0.000
#>   5:                     0.000                     0.000
#>  ---                                                    
#> 720:                     0.000                     0.000
#> 721:                     0.000                     0.000
#> 722:                     0.000                     0.000
#> 723:                     0.000                     0.000
#> 724:                     0.000                     0.000
#>      221012_DGM_MB1588_3_1_395 221012_DGM_MB1588_3_2_396
#>                          <num>                     <num>
#>   1:                     0.000                     0.000
#>   2:                     0.000                     0.000
#>   3:                     0.000                     0.000
#>   4:                     0.000                     0.000
#>   5:                     0.000                     0.000
#>  ---                                                    
#> 720:                     0.000                     0.000
#> 721:                     0.000                     0.000
#> 722:                     0.000                     0.000
#> 723:                  6487.566                  5098.135
#> 724:                202552.359                148361.500
#>      221012_DGM_MB1588_3_3_397 221012_DGM_MB1589_4_1_398
#>                          <num>                     <num>
#>   1:                     0.000                         0
#>   2:                     0.000                         0
#>   3:                     0.000                         0
#>   4:                     0.000                         0
#>   5:                     0.000                         0
#>  ---                                                    
#> 720:                     0.000                         0
#> 721:                     0.000                         0
#> 722:                     0.000                         0
#> 723:                  7221.981                         0
#> 724:                155704.094                         0
#>      221012_DGM_MB1589_4_2_399 221012_DGM_MB1589_4_3_400
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1590_5_1_401 221012_DGM_MB1590_5_2_402
#>                          <num>                     <num>
#>   1:                     0.000                     0.000
#>   2:                     0.000                     0.000
#>   3:                     0.000                     0.000
#>   4:                     0.000                     0.000
#>   5:                  1977.438                  1953.256
#>  ---                                                    
#> 720:                  5517.501                  4006.738
#> 721:                 17435.523                 17933.676
#> 722:                     0.000                     0.000
#> 723:                     0.000                     0.000
#> 724:                     0.000                     0.000
#>      221012_DGM_MB1590_5_3_403 221012_DGM_MB1591_6_1_407
#>                          <num>                     <num>
#>   1:                     0.000                         0
#>   2:                     0.000                         0
#>   3:                     0.000                         0
#>   4:                     0.000                         0
#>   5:                  2209.528                         0
#>  ---                                                    
#> 720:                  4882.215                         0
#> 721:                 18599.828                         0
#> 722:                     0.000                         0
#> 723:                     0.000                         0
#> 724:                     0.000                         0
#>      221012_DGM_MB1591_6_2_408 221012_DGM_MB1591_6_3_409
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1592_7_1_410 221012_DGM_MB1592_7_2_411
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1592_7_3_412 221012_DGM_MB1593_8_1_413
#>                          <num>                     <num>
#>   1:                         0                     0.000
#>   2:                         0                     0.000
#>   3:                         0                     0.000
#>   4:                         0                  3522.783
#>   5:                         0                     0.000
#>  ---                                                    
#> 720:                         0                     0.000
#> 721:                         0                     0.000
#> 722:                         0                 14306.895
#> 723:                         0                     0.000
#> 724:                         0                     0.000
#>      221012_DGM_MB1593_8_2_414 221012_DGM_MB1593_8_3_415
#>                          <num>                     <num>
#>   1:                      0.00                     0.000
#>   2:                      0.00                     0.000
#>   3:                      0.00                     0.000
#>   4:                   5010.69                  3815.988
#>   5:                      0.00                     0.000
#>  ---                                                    
#> 720:                      0.00                     0.000
#> 721:                      0.00                     0.000
#> 722:                  10564.97                 14566.125
#> 723:                      0.00                     0.000
#> 724:                      0.00                     0.000
#>      221012_DGM_MB1594_9_1_416 221012_DGM_MB1594_9_2_417
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1594_9_3_418 221012_DGM_MB1595_10_1_422
#>                          <num>                      <num>
#>   1:                         0                   9122.671
#>   2:                         0                  35174.156
#>   3:                         0                      0.000
#>   4:                         0                   1316.851
#>   5:                         0                      0.000
#>  ---                                                     
#> 720:                         0                      0.000
#> 721:                         0                      0.000
#> 722:                         0                      0.000
#> 723:                         0                      0.000
#> 724:                         0                      0.000
#>      221012_DGM_MB1595_10_2_423 221012_DGM_MB1595_10_3_424
#>                           <num>                      <num>
#>   1:                   9405.939                  10668.905
#>   2:                  32907.184                  30415.164
#>   3:                      0.000                      0.000
#>   4:                   1761.997                   1803.242
#>   5:                      0.000                      0.000
#>  ---                                                      
#> 720:                      0.000                      0.000
#> 721:                      0.000                      0.000
#> 722:                      0.000                      0.000
#> 723:                      0.000                      0.000
#> 724:                      0.000                      0.000
#>      221012_DGM_MB1597_11_1_425 221012_DGM_MB1597_11_2_426
#>                           <num>                      <num>
#>   1:                      0.000                       0.00
#>   2:                      0.000                       0.00
#>   3:                      0.000                       0.00
#>   4:                      0.000                       0.00
#>   5:                      0.000                       0.00
#>  ---                                                      
#> 720:                      0.000                       0.00
#> 721:                      0.000                       0.00
#> 722:                   7123.614                    9147.79
#> 723:                      0.000                       0.00
#> 724:                      0.000                       0.00
#>      221012_DGM_MB1597_11_3_427 221012_DGM_MB1598_12_1_428
#>                           <num>                      <num>
#>   1:                      0.000                          0
#>   2:                      0.000                          0
#>   3:                      0.000                          0
#>   4:                      0.000                          0
#>   5:                      0.000                          0
#>  ---                                                      
#> 720:                      0.000                          0
#> 721:                      0.000                          0
#> 722:                   9354.261                          0
#> 723:                      0.000                          0
#> 724:                      0.000                          0
#>      221012_DGM_MB1598_12_2_429 221012_DGM_MB1598_12_3_430
#>                           <num>                      <num>
#>   1:                          0                          0
#>   2:                          0                          0
#>   3:                          0                          0
#>   4:                          0                          0
#>   5:                          0                          0
#>  ---                                                      
#> 720:                          0                          0
#> 721:                          0                          0
#> 722:                          0                          0
#> 723:                          0                          0
#> 724:                          0                          0
#>      221012_DGM_MB1599_13_1_431 221012_DGM_MB1599_13_2_432
#>                           <num>                      <num>
#>   1:                   10114.79                   10594.81
#>   2:                       0.00                       0.00
#>   3:                   15386.47                   15706.29
#>   4:                       0.00                       0.00
#>   5:                       0.00                       0.00
#>  ---                                                      
#> 720:                       0.00                       0.00
#> 721:                       0.00                       0.00
#> 722:                       0.00                       0.00
#> 723:                       0.00                       0.00
#> 724:                       0.00                       0.00
#>      221012_DGM_MB1599_13_3_433    cor
#>                           <num> <lgcl>
#>   1:                   10425.28   TRUE
#>   2:                       0.00   TRUE
#>   3:                   16543.69   TRUE
#>   4:                       0.00   TRUE
#>   5:                       0.00   TRUE
#>  ---                                  
#> 720:                       0.00   TRUE
#> 721:                       0.00   TRUE
#> 722:                       0.00   TRUE
#> 723:                       0.00   TRUE
#> 724:                       0.00   TRUE

matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 filtered_data, 10, 6)
#> [1] "Reading: /home/runner/work/_temp/Library/mums2/extdata/botryllus_v2.gnps.mgf ..."
#> Computing                                                    | 0%  ETA: -...Computing ■                                                  | 2%  ETA: ...Computing ■■                                                 | 4%  ETA: ...Computing ■■■                                                | 6%  ETA: ...Computing ■■■■                                               | 8%  ETA: ...Computing ■■■■■                                              | 10%  ETA: ...Computing ■■■■■■                                             | 12%  ETA: ...Computing ■■■■■■■                                            | 14%  ETA: ...Computing ■■■■■■■■                                           | 16%  ETA: ...Computing ■■■■■■■■■                                          | 18%  ETA: ...Computing ■■■■■■■■■■                                         | 20%  ETA: ...Computing ■■■■■■■■■■■                                        | 22%  ETA: ...Computing ■■■■■■■■■■■■                                       | 24%  ETA: ...Computing ■■■■■■■■■■■■■                                      | 26%  ETA: ...Computing ■■■■■■■■■■■■■■                                     | 28%  ETA: ...Computing ■■■■■■■■■■■■■■■                                    | 30%  ETA: ...Computing ■■■■■■■■■■■■■■■■                                   | 32%  ETA: ...Computing ■■■■■■■■■■■■■■■■■                                  | 34%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■                                 | 36%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■                                | 38%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■                               | 40%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■                              | 42%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■                             | 44%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■                            | 46%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■                           | 48%  ETA: 1s ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■                          | 50%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■                         | 52%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■                        | 54%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■                       | 56%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                      | 58%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                     | 60%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                    | 62%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                   | 64%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                  | 66%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                 | 68%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                | 70%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■               | 72%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■              | 74%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■             | 76%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            | 78%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■           | 80%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■          | 82%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■         | 84%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■        | 86%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■       | 88%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■      | 90%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     | 92%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    | 94%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   | 96%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  | 98%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ | 100%  ETA: ...
#> [1] "41/724 peaks have an MS2 spectra."

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = modified_cosine_params(0.5), min_peaks = 0)
#> Computing                                                    | 0%  ETA: -...Computing ■                                                  | 2%  ETA: ...Computing ■■                                                 | 4%  ETA: ...Computing ■■■                                                | 7%  ETA: ...Computing ■■■■                                               | 9%  ETA: ...Computing ■■■■■■                                             | 12%  ETA: ...Computing ■■■■■■■                                            | 14%  ETA: ...Computing ■■■■■■■■                                           | 17%  ETA: ...Computing ■■■■■■■■■                                          | 19%  ETA: ...Computing ■■■■■■■■■■                                         | 21%  ETA: ...Computing ■■■■■■■■■■■■                                       | 24%  ETA: ...Computing ■■■■■■■■■■■■■                                      | 26%  ETA: ...Computing ■■■■■■■■■■■■■■                                     | 29%  ETA: ...Computing ■■■■■■■■■■■■■■■                                    | 31%  ETA: ...Computing ■■■■■■■■■■■■■■■■■                                  | 34%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■                                 | 36%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■                                | 39%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■                               | 41%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■                              | 43%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■                            | 46%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■                           | 48%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■                          | 51%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■                         | 53%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■                       | 56%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                      | 58%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                     | 60%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                    | 63%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                   | 65%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                 | 68%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                | 70%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■               | 73%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■              | 75%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            | 78%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■           | 80%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■          | 82%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■         | 85%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■        | 87%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■      | 90%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     | 92%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    | 95%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   | 97%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ | 100%  ETA: ...

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data, cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)
print(community_object)
#>                                omu1     omu2     omu3     omu4      omu5
#> 221012_DGM_Blank4_1_2_435       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank4_1_1_434       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1599_13_3_433 532663.7 285459.3 93695.73 31914.63     0.000
#> 221012_DGM_MB1599_13_1_431 491299.9 294798.5 92544.75 33688.02     0.000
#> 221012_DGM_MB1598_12_3_430      0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1598_12_1_428      0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1597_11_2_426 436238.6 277162.2 76928.43 26620.33     0.000
#> 221012_DGM_MB1595_10_3_424      0.0      0.0     0.00     0.00  8245.298
#> 221012_DGM_MB1595_10_2_423      0.0      0.0     0.00     0.00  8137.556
#> 221012_DGM_MB1597_11_3_427 538377.6 301881.6 82415.30 34167.55     0.000
#> 221012_DGM_MB1595_10_1_422      0.0      0.0     0.00     0.00 10280.282
#> 221012_DGM_Blank2_1_1_404       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank3_1_1_419       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1590_5_3_403       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank1_1_2_391       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1599_13_2_432 521400.8 315335.6 84531.67 26558.03     0.000
#> 221012_DGM_MB1590_5_2_402       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1597_11_1_425 466006.3 280857.9 83721.55 27610.95     0.000
#> 221012_DGM_Blank1_1_1_390       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1590_5_1_401       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1588_3_3_397       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1588_3_1_395       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank1_1_3_392       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1592_7_2_411       0.0      0.0     0.00     0.00 37276.324
#> 221012_DGM_MB1591_6_1_407       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1589_4_1_398       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank4_1_3_436       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank3_1_3_421       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1592_7_3_412       0.0      0.0     0.00     0.00 30775.895
#> 221012_DGM_MB1593_8_2_414       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1589_4_3_400       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1598_12_2_429      0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1589_4_2_399       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank2_1_2_405       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1591_6_2_408       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_Blank2_1_3_406       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1593_8_1_413       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1591_6_3_409       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1592_7_1_410       0.0      0.0     0.00     0.00 28894.629
#> 221012_DGM_MB1588_3_2_396       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1593_8_3_415       0.0      0.0     0.00     0.00     0.000
#> 221012_DGM_MB1594_9_1_416       0.0      0.0     0.00     0.00 18139.420
#> 221012_DGM_MB1594_9_2_417       0.0      0.0     0.00     0.00 21133.680
#> 221012_DGM_MB1594_9_3_418       0.0      0.0     0.00     0.00 26365.955
#> 221012_DGM_Blank3_1_2_420       0.0      0.0     0.00     0.00     0.000
#>                                omu6      omu7      omu8       omu9    omu10
#> 221012_DGM_Blank4_1_2_435      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_Blank4_1_1_434      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1599_13_3_433     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1599_13_1_431     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1598_12_3_430     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1598_12_1_428     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1597_11_2_426     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1595_10_3_424 28165.12 10029.085     0.000 128246.430     0.00
#> 221012_DGM_MB1595_10_2_423 20329.55  8753.449     0.000 126034.781     0.00
#> 221012_DGM_MB1597_11_3_427     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1595_10_1_422 24373.96 11221.259     0.000 116974.570     0.00
#> 221012_DGM_Blank2_1_1_404      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_Blank3_1_1_419      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1590_5_3_403      0.00 12424.777 16919.920      0.000     0.00
#> 221012_DGM_Blank1_1_2_391      0.00     0.000     0.000   2518.342     0.00
#> 221012_DGM_MB1599_13_2_432     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1590_5_2_402      0.00 10865.707 14397.848      0.000     0.00
#> 221012_DGM_MB1597_11_1_425     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_Blank1_1_1_390      0.00     0.000     0.000   3268.017     0.00
#> 221012_DGM_MB1590_5_1_401      0.00  9988.660 14096.659      0.000     0.00
#> 221012_DGM_MB1588_3_3_397      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1588_3_1_395      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_Blank1_1_3_392      0.00     0.000     0.000   3069.915     0.00
#> 221012_DGM_MB1592_7_2_411  47936.48 43576.500 44611.074 397249.906     0.00
#> 221012_DGM_MB1591_6_1_407  17509.81  6812.046  7249.338  66394.781     0.00
#> 221012_DGM_MB1589_4_1_398      0.00  9568.786     0.000      0.000 29497.71
#> 221012_DGM_Blank4_1_3_436      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_Blank3_1_3_421      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1592_7_3_412  40343.21 37458.512 38988.289 419085.594     0.00
#> 221012_DGM_MB1593_8_2_414  40522.22     0.000     0.000  51809.664     0.00
#> 221012_DGM_MB1589_4_3_400      0.00  8065.985     0.000      0.000 37300.27
#> 221012_DGM_MB1598_12_2_429     0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1589_4_2_399      0.00  9193.380     0.000      0.000 29970.38
#> 221012_DGM_Blank2_1_2_405      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1591_6_2_408  22342.92  8859.415  6991.487  69101.164     0.00
#> 221012_DGM_Blank2_1_3_406      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1593_8_1_413  41923.98     0.000     0.000  54509.211     0.00
#> 221012_DGM_MB1591_6_3_409  19378.68  8029.002  9578.860  71386.883     0.00
#> 221012_DGM_MB1592_7_1_410  35290.84 39380.055 40532.645 335408.469     0.00
#> 221012_DGM_MB1588_3_2_396      0.00     0.000     0.000      0.000     0.00
#> 221012_DGM_MB1593_8_3_415  46603.43     0.000     0.000  63907.711     0.00
#> 221012_DGM_MB1594_9_1_416      0.00 13206.461     0.000  58222.355     0.00
#> 221012_DGM_MB1594_9_2_417      0.00 16152.790     0.000  58101.988     0.00
#> 221012_DGM_MB1594_9_3_418      0.00 14692.003     0.000  59609.641     0.00
#> 221012_DGM_Blank3_1_2_420      0.00     0.000     0.000      0.000     0.00
#>                                 omu11    omu12    omu13    omu14    omu15
#> 221012_DGM_Blank4_1_2_435    79362.03     0.00      0.0    0.000     0.00
#> 221012_DGM_Blank4_1_1_434    96443.31     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1599_13_3_433       0.00     0.00 171391.2    0.000     0.00
#> 221012_DGM_MB1599_13_1_431       0.00     0.00 161515.5    0.000     0.00
#> 221012_DGM_MB1598_12_3_430       0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1598_12_1_428       0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1597_11_2_426  162597.30 62455.37 136136.8    0.000     0.00
#> 221012_DGM_MB1595_10_3_424 2690262.50     0.00      0.0 3514.396     0.00
#> 221012_DGM_MB1595_10_2_423 1851541.00     0.00      0.0 4089.383     0.00
#> 221012_DGM_MB1597_11_3_427  198054.44 71128.14 134072.4    0.000     0.00
#> 221012_DGM_MB1595_10_1_422 2323331.75     0.00      0.0 2795.664     0.00
#> 221012_DGM_Blank2_1_1_404   190885.30     0.00      0.0    0.000     0.00
#> 221012_DGM_Blank3_1_1_419        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1590_5_3_403  1236710.50     0.00      0.0    0.000     0.00
#> 221012_DGM_Blank1_1_2_391        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1599_13_2_432       0.00     0.00 174822.4    0.000     0.00
#> 221012_DGM_MB1590_5_2_402  1100348.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1597_11_1_425  205781.94 56863.18 147355.6    0.000     0.00
#> 221012_DGM_Blank1_1_1_390        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1590_5_1_401  1201459.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1588_3_3_397        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1588_3_1_395        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_Blank1_1_3_392        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1592_7_2_411  4551395.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1591_6_1_407   983453.75     0.00      0.0 6217.721     0.00
#> 221012_DGM_MB1589_4_1_398  2451402.00     0.00      0.0    0.000     0.00
#> 221012_DGM_Blank4_1_3_436   103831.11     0.00      0.0    0.000     0.00
#> 221012_DGM_Blank3_1_3_421        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1592_7_3_412  4120859.25     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1593_8_2_414  1310794.62     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1589_4_3_400  2479043.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1598_12_2_429       0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1589_4_2_399  2333548.25     0.00      0.0    0.000     0.00
#> 221012_DGM_Blank2_1_2_405   173035.73     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1591_6_2_408  1142081.00     0.00      0.0 6558.389     0.00
#> 221012_DGM_Blank2_1_3_406   163500.38     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1593_8_1_413  1419435.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1591_6_3_409  1222753.00     0.00      0.0 6176.717     0.00
#> 221012_DGM_MB1592_7_1_410  4052082.75     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1588_3_2_396        0.00     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1593_8_3_415  1379293.12     0.00      0.0    0.000     0.00
#> 221012_DGM_MB1594_9_1_416   434276.91     0.00      0.0    0.000 17008.70
#> 221012_DGM_MB1594_9_2_417   547637.06     0.00      0.0    0.000 18944.96
#> 221012_DGM_MB1594_9_3_418   612912.94     0.00      0.0    0.000 15940.30
#> 221012_DGM_Blank3_1_2_420        0.00     0.00      0.0    0.000     0.00
#>                               omu16     omu17    omu18    omu19    omu20
#> 221012_DGM_Blank4_1_2_435     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_Blank4_1_1_434     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1599_13_3_433    0.000  5882.270 5971.201 66153.28 330217.0
#> 221012_DGM_MB1599_13_1_431    0.000  5900.368 5931.985 63568.83 261554.0
#> 221012_DGM_MB1598_12_3_430 1799.620     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1598_12_1_428 1494.314     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1597_11_2_426    0.000     0.000    0.000 87355.89 331224.4
#> 221012_DGM_MB1595_10_3_424    0.000  7048.427    0.000     0.00      0.0
#> 221012_DGM_MB1595_10_2_423    0.000  7128.086    0.000     0.00      0.0
#> 221012_DGM_MB1597_11_3_427    0.000     0.000    0.000 92277.95 316056.3
#> 221012_DGM_MB1595_10_1_422    0.000  7000.022    0.000     0.00      0.0
#> 221012_DGM_Blank2_1_1_404     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_Blank3_1_1_419     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1590_5_3_403  6080.775     0.000    0.000     0.00      0.0
#> 221012_DGM_Blank1_1_2_391     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1599_13_2_432    0.000  7298.647 6084.393 68630.12 258658.0
#> 221012_DGM_MB1590_5_2_402  5470.364     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1597_11_1_425    0.000     0.000    0.000 90921.17 329490.4
#> 221012_DGM_Blank1_1_1_390     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1590_5_1_401  4639.385     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1588_3_3_397     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1588_3_1_395     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_Blank1_1_3_392     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1592_7_2_411     0.000 28304.791    0.000     0.00      0.0
#> 221012_DGM_MB1591_6_1_407     0.000  6593.535    0.000     0.00      0.0
#> 221012_DGM_MB1589_4_1_398     0.000  5436.953    0.000     0.00      0.0
#> 221012_DGM_Blank4_1_3_436     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_Blank3_1_3_421     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1592_7_3_412     0.000 30503.092    0.000     0.00      0.0
#> 221012_DGM_MB1593_8_2_414     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1589_4_3_400     0.000  6279.896    0.000     0.00      0.0
#> 221012_DGM_MB1598_12_2_429 1690.556     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1589_4_2_399     0.000  5414.722    0.000     0.00      0.0
#> 221012_DGM_Blank2_1_2_405     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1591_6_2_408     0.000  7007.155    0.000     0.00      0.0
#> 221012_DGM_Blank2_1_3_406     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1593_8_1_413     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1591_6_3_409     0.000  5987.346    0.000     0.00      0.0
#> 221012_DGM_MB1592_7_1_410     0.000 25797.861    0.000     0.00      0.0
#> 221012_DGM_MB1588_3_2_396     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1593_8_3_415     0.000     0.000    0.000     0.00      0.0
#> 221012_DGM_MB1594_9_1_416     0.000 12352.333    0.000     0.00      0.0
#> 221012_DGM_MB1594_9_2_417     0.000  9627.950    0.000     0.00      0.0
#> 221012_DGM_MB1594_9_3_418     0.000 12442.386    0.000     0.00      0.0
#> 221012_DGM_Blank3_1_2_420     0.000     0.000    0.000     0.00      0.0
#>                               omu21       omu22       omu23     omu24    omu25
#> 221012_DGM_Blank4_1_2_435       0.0    5036.731       0.000     0.000     0.00
#> 221012_DGM_Blank4_1_1_434       0.0    7314.682       0.000     0.000     0.00
#> 221012_DGM_MB1599_13_3_433 327624.5  688076.750  592729.375     0.000     0.00
#> 221012_DGM_MB1599_13_1_431 318160.8  702697.938  545031.938     0.000     0.00
#> 221012_DGM_MB1598_12_3_430      0.0       0.000       0.000     0.000     0.00
#> 221012_DGM_MB1598_12_1_428      0.0       0.000       0.000     0.000     0.00
#> 221012_DGM_MB1597_11_2_426 359302.4  313380.125  272656.125     0.000 15584.79
#> 221012_DGM_MB1595_10_3_424      0.0 1050425.125  765332.438 16443.771     0.00
#> 221012_DGM_MB1595_10_2_423      0.0  983244.375  702855.750 12622.469     0.00
#> 221012_DGM_MB1597_11_3_427 366588.2  407948.781  326057.125     0.000 13822.27
#> 221012_DGM_MB1595_10_1_422      0.0  901533.562  724884.875 13342.501     0.00
#> 221012_DGM_Blank2_1_1_404       0.0    7136.955    9038.840     0.000     0.00
#> 221012_DGM_Blank3_1_1_419       0.0       0.000    5843.835     0.000     0.00
#> 221012_DGM_MB1590_5_3_403       0.0  647979.750  510211.312     0.000     0.00
#> 221012_DGM_Blank1_1_2_391       0.0    8501.517       0.000     0.000     0.00
#> 221012_DGM_MB1599_13_2_432 333043.1  687898.312  543989.312     0.000     0.00
#> 221012_DGM_MB1590_5_2_402       0.0  540991.438  417547.625     0.000     0.00
#> 221012_DGM_MB1597_11_1_425 377861.5  359382.875  291653.656     0.000 14754.62
#> 221012_DGM_Blank1_1_1_390       0.0    6896.363       0.000     0.000     0.00
#> 221012_DGM_MB1590_5_1_401       0.0  519021.938  420380.281     0.000     0.00
#> 221012_DGM_MB1588_3_3_397       0.0  391237.344  303101.938     0.000     0.00
#> 221012_DGM_MB1588_3_1_395       0.0  369823.969  301194.625     0.000     0.00
#> 221012_DGM_Blank1_1_3_392       0.0    6408.753       0.000     0.000     0.00
#> 221012_DGM_MB1592_7_2_411       0.0       0.000 3454213.000 16514.592     0.00
#> 221012_DGM_MB1591_6_1_407       0.0  486336.000  411362.375  4895.993     0.00
#> 221012_DGM_MB1589_4_1_398       0.0  339129.094  272703.281     0.000 13216.81
#> 221012_DGM_Blank4_1_3_436       0.0    6352.350       0.000     0.000     0.00
#> 221012_DGM_Blank3_1_3_421       0.0       0.000    6268.979     0.000     0.00
#> 221012_DGM_MB1592_7_3_412       0.0       0.000 3075383.500 15145.320     0.00
#> 221012_DGM_MB1593_8_2_414       0.0  682367.875  493001.438  8042.770     0.00
#> 221012_DGM_MB1589_4_3_400       0.0  373199.500  266208.969     0.000 13296.45
#> 221012_DGM_MB1598_12_2_429      0.0       0.000       0.000     0.000     0.00
#> 221012_DGM_MB1589_4_2_399       0.0  357438.312  301143.594     0.000 10507.92
#> 221012_DGM_Blank2_1_2_405       0.0    6563.983    8545.672     0.000     0.00
#> 221012_DGM_MB1591_6_2_408       0.0  549781.375  423094.375  4617.321     0.00
#> 221012_DGM_Blank2_1_3_406       0.0    9176.102    7382.974     0.000     0.00
#> 221012_DGM_MB1593_8_1_413       0.0  630282.625  494232.812  7844.511     0.00
#> 221012_DGM_MB1591_6_3_409       0.0  496350.219  380114.969  5571.872     0.00
#> 221012_DGM_MB1592_7_1_410       0.0       0.000 3592727.500 15537.005     0.00
#> 221012_DGM_MB1588_3_2_396       0.0  394413.375  319799.531     0.000     0.00
#> 221012_DGM_MB1593_8_3_415       0.0  699364.438  492117.500  7347.139     0.00
#> 221012_DGM_MB1594_9_1_416       0.0  682223.375  474316.281  5060.728     0.00
#> 221012_DGM_MB1594_9_2_417       0.0  732523.000  554592.062  4013.373     0.00
#> 221012_DGM_MB1594_9_3_418       0.0  730286.625  567810.812  3740.446     0.00
#> 221012_DGM_Blank3_1_2_420       0.0       0.000    5776.215     0.000     0.00
#>                               omu26     omu27    omu28   omu29     omu30
#> 221012_DGM_Blank4_1_2_435     0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank4_1_1_434     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1599_13_3_433    0.000  4679.062     0.00 1363357  89183.57
#> 221012_DGM_MB1599_13_1_431    0.000  4224.674     0.00 1268357  74257.58
#> 221012_DGM_MB1598_12_3_430    0.000  2997.880 30422.23 3029606 177931.75
#> 221012_DGM_MB1598_12_1_428    0.000  2363.108 31735.44 2968034 181067.69
#> 221012_DGM_MB1597_11_2_426    0.000     0.000 28335.28 1373807  91922.39
#> 221012_DGM_MB1595_10_3_424    0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1595_10_2_423    0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1597_11_3_427    0.000     0.000 20153.58 1360959  89106.45
#> 221012_DGM_MB1595_10_1_422    0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank2_1_1_404     0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank3_1_1_419     0.000  2605.679     0.00       0      0.00
#> 221012_DGM_MB1590_5_3_403  2889.123     0.000     0.00       0      0.00
#> 221012_DGM_Blank1_1_2_391     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1599_13_2_432    0.000  3529.262     0.00 1428312 109160.04
#> 221012_DGM_MB1590_5_2_402  2854.606     0.000     0.00       0      0.00
#> 221012_DGM_MB1597_11_1_425    0.000     0.000 25869.93 1486557  97051.70
#> 221012_DGM_Blank1_1_1_390     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1590_5_1_401  3090.998     0.000     0.00       0      0.00
#> 221012_DGM_MB1588_3_3_397     0.000 52264.266     0.00       0      0.00
#> 221012_DGM_MB1588_3_1_395     0.000 67501.492     0.00       0      0.00
#> 221012_DGM_Blank1_1_3_392     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1592_7_2_411     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1591_6_1_407     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1589_4_1_398     0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank4_1_3_436     0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank3_1_3_421     0.000  2617.444     0.00       0      0.00
#> 221012_DGM_MB1592_7_3_412     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1593_8_2_414     0.000 13656.555     0.00       0      0.00
#> 221012_DGM_MB1589_4_3_400     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1598_12_2_429    0.000  2409.298 28345.11 2868668 173745.05
#> 221012_DGM_MB1589_4_2_399     0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank2_1_2_405     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1591_6_2_408     0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank2_1_3_406     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1593_8_1_413     0.000 12638.061     0.00       0      0.00
#> 221012_DGM_MB1591_6_3_409     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1592_7_1_410     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1588_3_2_396     0.000 67074.508     0.00       0      0.00
#> 221012_DGM_MB1593_8_3_415     0.000 16502.172     0.00       0      0.00
#> 221012_DGM_MB1594_9_1_416     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1594_9_2_417     0.000     0.000     0.00       0      0.00
#> 221012_DGM_MB1594_9_3_418     0.000     0.000     0.00       0      0.00
#> 221012_DGM_Blank3_1_2_420     0.000  1984.634     0.00       0      0.00
#>                                omu31     omu32       omu33      omu34
#> 221012_DGM_Blank4_1_2_435      0.000     0.000   24921.734      0.000
#> 221012_DGM_Blank4_1_1_434      0.000     0.000   28253.215      0.000
#> 221012_DGM_MB1599_13_3_433  1849.604 14510.130       0.000      0.000
#> 221012_DGM_MB1599_13_1_431  1943.195 13183.521       0.000      0.000
#> 221012_DGM_MB1598_12_3_430     0.000 26456.867       0.000      0.000
#> 221012_DGM_MB1598_12_1_428     0.000 25710.537       0.000      0.000
#> 221012_DGM_MB1597_11_2_426 10128.913  8540.696 3183455.500  55521.062
#> 221012_DGM_MB1595_10_3_424  4580.583     0.000       0.000   8202.558
#> 221012_DGM_MB1595_10_2_423  4746.543     0.000       0.000   7321.925
#> 221012_DGM_MB1597_11_3_427 10662.664  9564.402 3746430.562  77625.391
#> 221012_DGM_MB1595_10_1_422  3605.796     0.000       0.000   7153.493
#> 221012_DGM_Blank2_1_1_404      0.000     0.000   22563.162      0.000
#> 221012_DGM_Blank3_1_1_419      0.000     0.000       0.000      0.000
#> 221012_DGM_MB1590_5_3_403      0.000     0.000   68767.266  48398.961
#> 221012_DGM_Blank1_1_2_391      0.000     0.000    4341.520      0.000
#> 221012_DGM_MB1599_13_2_432  1729.792 16300.138       0.000      0.000
#> 221012_DGM_MB1590_5_2_402      0.000     0.000   49289.641  36873.129
#> 221012_DGM_MB1597_11_1_425 10208.766 11529.541 3373673.625  62995.742
#> 221012_DGM_Blank1_1_1_390      0.000     0.000    4357.495      0.000
#> 221012_DGM_MB1590_5_1_401      0.000     0.000   55115.668  37303.871
#> 221012_DGM_MB1588_3_3_397      0.000     0.000       0.000      0.000
#> 221012_DGM_MB1588_3_1_395      0.000     0.000       0.000      0.000
#> 221012_DGM_Blank1_1_3_392      0.000     0.000    4780.165      0.000
#> 221012_DGM_MB1592_7_2_411      0.000     0.000  384481.188 142510.484
#> 221012_DGM_MB1591_6_1_407      0.000     0.000   75051.812  61458.453
#> 221012_DGM_MB1589_4_1_398      0.000     0.000  174472.836 118462.000
#> 221012_DGM_Blank4_1_3_436      0.000     0.000   27743.361      0.000
#> 221012_DGM_Blank3_1_3_421      0.000     0.000       0.000      0.000
#> 221012_DGM_MB1592_7_3_412      0.000     0.000  331147.219 133667.234
#> 221012_DGM_MB1593_8_2_414   1579.983     0.000  184789.355      0.000
#> 221012_DGM_MB1589_4_3_400      0.000     0.000  187464.242  90262.664
#> 221012_DGM_MB1598_12_2_429     0.000 23067.719       0.000      0.000
#> 221012_DGM_MB1589_4_2_399      0.000     0.000  188013.328  87342.336
#> 221012_DGM_Blank2_1_2_405      0.000     0.000   23553.248      0.000
#> 221012_DGM_MB1591_6_2_408      0.000     0.000   84494.961  54561.567
#> 221012_DGM_Blank2_1_3_406      0.000     0.000   25887.674      0.000
#> 221012_DGM_MB1593_8_1_413   1560.431     0.000  187224.180      0.000
#> 221012_DGM_MB1591_6_3_409      0.000     0.000   81532.688  68659.176
#> 221012_DGM_MB1592_7_1_410      0.000     0.000  333042.750 105408.805
#> 221012_DGM_MB1588_3_2_396      0.000     0.000       0.000      0.000
#> 221012_DGM_MB1593_8_3_415   1416.351     0.000  193376.125      0.000
#> 221012_DGM_MB1594_9_1_416   4410.679     0.000  289688.375 603660.469
#> 221012_DGM_MB1594_9_2_417   4318.887     0.000  361953.633 656385.915
#> 221012_DGM_MB1594_9_3_418   5275.683     0.000  351155.523 721004.643
#> 221012_DGM_Blank3_1_2_420      0.000     0.000       0.000      0.000
#>                                  omu35      omu36
#> 221012_DGM_Blank4_1_2_435       0.0000      0.000
#> 221012_DGM_Blank4_1_1_434       0.0000      0.000
#> 221012_DGM_MB1599_13_3_433 477337.2031  84269.012
#> 221012_DGM_MB1599_13_1_431 420071.2501 113462.480
#> 221012_DGM_MB1598_12_3_430      0.0000      0.000
#> 221012_DGM_MB1598_12_1_428      0.0000      0.000
#> 221012_DGM_MB1597_11_2_426 433094.1094   7994.746
#> 221012_DGM_MB1595_10_3_424      0.0000      0.000
#> 221012_DGM_MB1595_10_2_423      0.0000      0.000
#> 221012_DGM_MB1597_11_3_427 427183.0000  10357.960
#> 221012_DGM_MB1595_10_1_422      0.0000      0.000
#> 221012_DGM_Blank2_1_1_404       0.0000      0.000
#> 221012_DGM_Blank3_1_1_419       0.0000      0.000
#> 221012_DGM_MB1590_5_3_403       0.0000  51747.383
#> 221012_DGM_Blank1_1_2_391       0.0000      0.000
#> 221012_DGM_MB1599_13_2_432 423367.3594  93951.492
#> 221012_DGM_MB1590_5_2_402       0.0000  37704.405
#> 221012_DGM_MB1597_11_1_425 445653.8438  11945.618
#> 221012_DGM_Blank1_1_1_390       0.0000      0.000
#> 221012_DGM_MB1590_5_1_401       0.0000  46584.157
#> 221012_DGM_MB1588_3_3_397       0.0000      0.000
#> 221012_DGM_MB1588_3_1_395       0.0000      0.000
#> 221012_DGM_Blank1_1_3_392       0.0000      0.000
#> 221012_DGM_MB1592_7_2_411     964.2026  19686.236
#> 221012_DGM_MB1591_6_1_407       0.0000  18517.891
#> 221012_DGM_MB1589_4_1_398       0.0000  22551.477
#> 221012_DGM_Blank4_1_3_436       0.0000      0.000
#> 221012_DGM_Blank3_1_3_421       0.0000      0.000
#> 221012_DGM_MB1592_7_3_412     812.7355  18892.471
#> 221012_DGM_MB1593_8_2_414       0.0000      0.000
#> 221012_DGM_MB1589_4_3_400       0.0000  26448.909
#> 221012_DGM_MB1598_12_2_429      0.0000      0.000
#> 221012_DGM_MB1589_4_2_399       0.0000  21068.432
#> 221012_DGM_Blank2_1_2_405       0.0000      0.000
#> 221012_DGM_MB1591_6_2_408       0.0000  18015.770
#> 221012_DGM_Blank2_1_3_406       0.0000      0.000
#> 221012_DGM_MB1593_8_1_413       0.0000      0.000
#> 221012_DGM_MB1591_6_3_409       0.0000  18514.527
#> 221012_DGM_MB1592_7_1_410     804.7273  14075.879
#> 221012_DGM_MB1588_3_2_396       0.0000      0.000
#> 221012_DGM_MB1593_8_3_415       0.0000      0.000
#> 221012_DGM_MB1594_9_1_416       0.0000      0.000
#> 221012_DGM_MB1594_9_2_417       0.0000      0.000
#> 221012_DGM_MB1594_9_3_418       0.0000      0.000
#> 221012_DGM_Blank3_1_2_420       0.0000      0.000