Rarefy MS1 Feature Table

rarefy_ms() performs a single subsampling of MS1 features in sample. Feature intensities are subsampled to the supplied size and accounts for intensity thresholds due to machine limits and background noise. Specifically, features whose abundance falls below the threshold after rarefying are removed. This allows for accurate representation of samples at different dilutions regardless of the desired submsampling size.

Usage

rarefy_ms(
  community_object,
  size,
  threshold,
  number_of_threads = detectCores(),
  seed = 123
)

Arguments

community_object

A community_object

size

The desired total sample intensity to subsample to.

threshold

The individual feature threshold. Each subsampled feature must be >= this value to be retained.

number_of_threads

the amount of threads you want the calculation to use.

seed

the RNG (random number generator) seed you would like to use.

Value

A external_pointer that references a community matrix of rarefied feature intensities.

returns a matrix object that contains your rarefied data.

Examples

data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   meta_data =
                   mums2::mums2_example("meta_data_boryillus.csv"),
                   format = "None")
#> If peak table has corrupted compound names they will be converted to
#>       utf-8 and if there are any commas, they will be converted to periods(.).

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())
#> ℹ Checking 1500 peaks for mispicked peaks.
#> ℹ Argument merge_peaks is: TRUE. Merging mispicked peaks with method sum.
#> ✔ 50 ions failed the mispicked filter, 1450 ions remain.
#> ℹ Parsing 1450 peaks for replicability across technical replicates.
#> ✔ 329 ions failed the cv_filter filter, 1121 ions remain.
#> ℹ Parsing 1121 peaks based on the sample group: Blanks.
#> ℹ Argument remove_ions is: TRUE.Removing peaks from Blanks.
#> ✔ 342 ions failed the Blanks filter, 779 ions remain.
#> ℹ Parsing 779 peaks for insource ions.
#> ✔ 55 ions failed the insource filter, 724 ions remain.

change_rt_to_seconds_or_minute(filtered_data, "minutes")
#> [1] "Changing rt values to minutes"
#> Key: <Compound, mz, kmd, RTINMINUTES>
#>                    Compound        mz     kmd RTINMINUTES
#>                      <char>     <num>   <num>       <num>
#>   1: 1000.65345 Da 418.99 s 1001.6607 0.66073   0.1163333
#>   2: 1002.23833 Da 440.75 s 1003.2456 0.24560   0.1225000
#>   3: 1002.57692 Da 414.74 s 1003.5842 0.58419   0.1151667
#>   4: 1004.19672 Da 603.96 s 1005.2040 0.20400   0.1678333
#>   5: 1004.23580 Da 548.28 s 1005.2431 0.24307   0.1523333
#>  ---                                                     
#> 720:  980.84769 Da 515.58 s  981.8550 0.85497   0.1431667
#> 721:  984.23693 Da 604.23 s  985.2442 0.24420   0.1678333
#> 722:  984.65546 Da 460.05 s  985.6627 0.66274   0.1278333
#> 723:  997.30596 Da 605.90 s  998.3132 0.31323   0.1683333
#> 724:  998.28767 Da 556.84 s  981.2843 0.28429   0.1546667
#>      221012_DGM_Blank1_1_1_390 221012_DGM_Blank1_1_2_391
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank1_1_3_392 221012_DGM_Blank2_1_1_404
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank2_1_2_405 221012_DGM_Blank2_1_3_406
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank3_1_1_419 221012_DGM_Blank3_1_2_420
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_Blank3_1_3_421 221012_DGM_Blank4_1_1_434
#>                          <num>                     <num>
#>   1:                         0                   1538.23
#>   2:                         0                      0.00
#>   3:                         0                      0.00
#>   4:                         0                      0.00
#>   5:                         0                      0.00
#>  ---                                                    
#> 720:                         0                      0.00
#> 721:                         0                      0.00
#> 722:                         0                      0.00
#> 723:                         0                      0.00
#> 724:                         0                      0.00
#>      221012_DGM_Blank4_1_2_435 221012_DGM_Blank4_1_3_436
#>                          <num>                     <num>
#>   1:                  1201.261                  1180.144
#>   2:                     0.000                     0.000
#>   3:                     0.000                     0.000
#>   4:                     0.000                     0.000
#>   5:                     0.000                     0.000
#>  ---                                                    
#> 720:                     0.000                     0.000
#> 721:                     0.000                     0.000
#> 722:                     0.000                     0.000
#> 723:                     0.000                     0.000
#> 724:                     0.000                     0.000
#>      221012_DGM_MB1588_3_1_395 221012_DGM_MB1588_3_2_396
#>                          <num>                     <num>
#>   1:                     0.000                     0.000
#>   2:                     0.000                     0.000
#>   3:                     0.000                     0.000
#>   4:                     0.000                     0.000
#>   5:                     0.000                     0.000
#>  ---                                                    
#> 720:                     0.000                     0.000
#> 721:                     0.000                     0.000
#> 722:                     0.000                     0.000
#> 723:                  6487.566                  5098.135
#> 724:                202552.359                148361.500
#>      221012_DGM_MB1588_3_3_397 221012_DGM_MB1589_4_1_398
#>                          <num>                     <num>
#>   1:                     0.000                         0
#>   2:                     0.000                         0
#>   3:                     0.000                         0
#>   4:                     0.000                         0
#>   5:                     0.000                         0
#>  ---                                                    
#> 720:                     0.000                         0
#> 721:                     0.000                         0
#> 722:                     0.000                         0
#> 723:                  7221.981                         0
#> 724:                155704.094                         0
#>      221012_DGM_MB1589_4_2_399 221012_DGM_MB1589_4_3_400
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1590_5_1_401 221012_DGM_MB1590_5_2_402
#>                          <num>                     <num>
#>   1:                     0.000                     0.000
#>   2:                     0.000                     0.000
#>   3:                     0.000                     0.000
#>   4:                     0.000                     0.000
#>   5:                  1977.438                  1953.256
#>  ---                                                    
#> 720:                  5517.501                  4006.738
#> 721:                 17435.523                 17933.676
#> 722:                     0.000                     0.000
#> 723:                     0.000                     0.000
#> 724:                     0.000                     0.000
#>      221012_DGM_MB1590_5_3_403 221012_DGM_MB1591_6_1_407
#>                          <num>                     <num>
#>   1:                     0.000                         0
#>   2:                     0.000                         0
#>   3:                     0.000                         0
#>   4:                     0.000                         0
#>   5:                  2209.528                         0
#>  ---                                                    
#> 720:                  4882.215                         0
#> 721:                 18599.828                         0
#> 722:                     0.000                         0
#> 723:                     0.000                         0
#> 724:                     0.000                         0
#>      221012_DGM_MB1591_6_2_408 221012_DGM_MB1591_6_3_409
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1592_7_1_410 221012_DGM_MB1592_7_2_411
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1592_7_3_412 221012_DGM_MB1593_8_1_413
#>                          <num>                     <num>
#>   1:                         0                     0.000
#>   2:                         0                     0.000
#>   3:                         0                     0.000
#>   4:                         0                  3522.783
#>   5:                         0                     0.000
#>  ---                                                    
#> 720:                         0                     0.000
#> 721:                         0                     0.000
#> 722:                         0                 14306.895
#> 723:                         0                     0.000
#> 724:                         0                     0.000
#>      221012_DGM_MB1593_8_2_414 221012_DGM_MB1593_8_3_415
#>                          <num>                     <num>
#>   1:                      0.00                     0.000
#>   2:                      0.00                     0.000
#>   3:                      0.00                     0.000
#>   4:                   5010.69                  3815.988
#>   5:                      0.00                     0.000
#>  ---                                                    
#> 720:                      0.00                     0.000
#> 721:                      0.00                     0.000
#> 722:                  10564.97                 14566.125
#> 723:                      0.00                     0.000
#> 724:                      0.00                     0.000
#>      221012_DGM_MB1594_9_1_416 221012_DGM_MB1594_9_2_417
#>                          <num>                     <num>
#>   1:                         0                         0
#>   2:                         0                         0
#>   3:                         0                         0
#>   4:                         0                         0
#>   5:                         0                         0
#>  ---                                                    
#> 720:                         0                         0
#> 721:                         0                         0
#> 722:                         0                         0
#> 723:                         0                         0
#> 724:                         0                         0
#>      221012_DGM_MB1594_9_3_418 221012_DGM_MB1595_10_1_422
#>                          <num>                      <num>
#>   1:                         0                   9122.671
#>   2:                         0                  35174.156
#>   3:                         0                      0.000
#>   4:                         0                   1316.851
#>   5:                         0                      0.000
#>  ---                                                     
#> 720:                         0                      0.000
#> 721:                         0                      0.000
#> 722:                         0                      0.000
#> 723:                         0                      0.000
#> 724:                         0                      0.000
#>      221012_DGM_MB1595_10_2_423 221012_DGM_MB1595_10_3_424
#>                           <num>                      <num>
#>   1:                   9405.939                  10668.905
#>   2:                  32907.184                  30415.164
#>   3:                      0.000                      0.000
#>   4:                   1761.997                   1803.242
#>   5:                      0.000                      0.000
#>  ---                                                      
#> 720:                      0.000                      0.000
#> 721:                      0.000                      0.000
#> 722:                      0.000                      0.000
#> 723:                      0.000                      0.000
#> 724:                      0.000                      0.000
#>      221012_DGM_MB1597_11_1_425 221012_DGM_MB1597_11_2_426
#>                           <num>                      <num>
#>   1:                      0.000                       0.00
#>   2:                      0.000                       0.00
#>   3:                      0.000                       0.00
#>   4:                      0.000                       0.00
#>   5:                      0.000                       0.00
#>  ---                                                      
#> 720:                      0.000                       0.00
#> 721:                      0.000                       0.00
#> 722:                   7123.614                    9147.79
#> 723:                      0.000                       0.00
#> 724:                      0.000                       0.00
#>      221012_DGM_MB1597_11_3_427 221012_DGM_MB1598_12_1_428
#>                           <num>                      <num>
#>   1:                      0.000                          0
#>   2:                      0.000                          0
#>   3:                      0.000                          0
#>   4:                      0.000                          0
#>   5:                      0.000                          0
#>  ---                                                      
#> 720:                      0.000                          0
#> 721:                      0.000                          0
#> 722:                   9354.261                          0
#> 723:                      0.000                          0
#> 724:                      0.000                          0
#>      221012_DGM_MB1598_12_2_429 221012_DGM_MB1598_12_3_430
#>                           <num>                      <num>
#>   1:                          0                          0
#>   2:                          0                          0
#>   3:                          0                          0
#>   4:                          0                          0
#>   5:                          0                          0
#>  ---                                                      
#> 720:                          0                          0
#> 721:                          0                          0
#> 722:                          0                          0
#> 723:                          0                          0
#> 724:                          0                          0
#>      221012_DGM_MB1599_13_1_431 221012_DGM_MB1599_13_2_432
#>                           <num>                      <num>
#>   1:                   10114.79                   10594.81
#>   2:                       0.00                       0.00
#>   3:                   15386.47                   15706.29
#>   4:                       0.00                       0.00
#>   5:                       0.00                       0.00
#>  ---                                                      
#> 720:                       0.00                       0.00
#> 721:                       0.00                       0.00
#> 722:                       0.00                       0.00
#> 723:                       0.00                       0.00
#> 724:                       0.00                       0.00
#>      221012_DGM_MB1599_13_3_433    cor
#>                           <num> <lgcl>
#>   1:                   10425.28   TRUE
#>   2:                       0.00   TRUE
#>   3:                   16543.69   TRUE
#>   4:                       0.00   TRUE
#>   5:                       0.00   TRUE
#>  ---                                  
#> 720:                       0.00   TRUE
#> 721:                       0.00   TRUE
#> 722:                       0.00   TRUE
#> 723:                       0.00   TRUE
#> 724:                       0.00   TRUE

matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 filtered_data, 10, 6)
#> [1] "Reading: /home/runner/work/_temp/Library/mums2/extdata/botryllus_v2.gnps.mgf ..."
#> Computing                                                    | 0%  ETA: -...Computing ■                                                  | 2%  ETA: ...Computing ■■                                                 | 4%  ETA: ...Computing ■■■                                                | 6%  ETA: ...Computing ■■■■                                               | 8%  ETA: ...Computing ■■■■■                                              | 10%  ETA: ...Computing ■■■■■■                                             | 12%  ETA: ...Computing ■■■■■■■                                            | 14%  ETA: ...Computing ■■■■■■■■                                           | 16%  ETA: ...Computing ■■■■■■■■■                                          | 18%  ETA: ...Computing ■■■■■■■■■■                                         | 20%  ETA: ...Computing ■■■■■■■■■■■                                        | 22%  ETA: ...Computing ■■■■■■■■■■■■                                       | 24%  ETA: ...Computing ■■■■■■■■■■■■■                                      | 26%  ETA: ...Computing ■■■■■■■■■■■■■■                                     | 28%  ETA: ...Computing ■■■■■■■■■■■■■■■                                    | 30%  ETA: ...Computing ■■■■■■■■■■■■■■■■                                   | 32%  ETA: ...Computing ■■■■■■■■■■■■■■■■■                                  | 34%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■                                 | 36%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■                                | 38%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■                               | 40%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■                              | 42%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■                             | 44%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■                            | 46%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■                           | 48%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■                          | 50%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■                         | 52%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■                        | 54%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■                       | 56%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                      | 58%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                     | 60%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                    | 62%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                   | 64%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                  | 66%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                 | 68%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                | 70%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■               | 72%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■              | 74%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■             | 76%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            | 78%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■           | 80%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■          | 82%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■         | 84%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■        | 86%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■       | 88%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■      | 90%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     | 92%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    | 94%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   | 96%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  | 98%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ | 100%  ETA: ...
#> [1] "41/724 peaks have an MS2 spectra."

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = modified_cosine_params(0.5), min_peaks = 0)
#> Computing                                                    | 0%  ETA: -...Computing ■                                                  | 2%  ETA: ...Computing ■■                                                 | 4%  ETA: ...Computing ■■■                                                | 7%  ETA: ...Computing ■■■■                                               | 9%  ETA: ...Computing ■■■■■■                                             | 12%  ETA: ...Computing ■■■■■■■                                            | 14%  ETA: ...Computing ■■■■■■■■                                           | 17%  ETA: ...Computing ■■■■■■■■■                                          | 19%  ETA: ...Computing ■■■■■■■■■■                                         | 21%  ETA: ...Computing ■■■■■■■■■■■■                                       | 24%  ETA: ...Computing ■■■■■■■■■■■■■                                      | 26%  ETA: ...Computing ■■■■■■■■■■■■■■                                     | 29%  ETA: ...Computing ■■■■■■■■■■■■■■■                                    | 31%  ETA: ...Computing ■■■■■■■■■■■■■■■■■                                  | 34%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■                                 | 36%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■                                | 39%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■                               | 41%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■                              | 43%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■                            | 46%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■                           | 48%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■                          | 51%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■                         | 53%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■                       | 56%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                      | 58%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                     | 60%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                    | 63%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                   | 65%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                 | 68%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                | 70%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■               | 73%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■              | 75%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            | 78%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■           | 80%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■          | 82%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■         | 85%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■        | 87%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■      | 90%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     | 92%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    | 95%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   | 97%  ETA: ...Computing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ | 100%  ETA: ...

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data, cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)
rarefy_ms(community_object, 4000, 100)
#>                            omu1 omu2 omu3 omu4 omu5 omu6 omu7 omu8 omu9 omu10
#> 221012_DGM_Blank4_1_2_435     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank4_1_1_434     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1599_13_3_433  472  246    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1599_13_1_431  414  254    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1598_12_3_430    0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1598_12_1_428    0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1597_11_2_426  265  134    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1595_10_3_424    0    0    0    0    0    0    0    0  110     0
#> 221012_DGM_MB1595_10_2_423    0    0    0    0    0    0    0    0  145     0
#> 221012_DGM_MB1597_11_3_427  275  142    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1595_10_1_422    0    0    0    0    0    0    0    0  132     0
#> 221012_DGM_Blank2_1_1_404     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank3_1_1_419     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1590_5_3_403     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank1_1_2_391     0    0    0    0    0    0    0    0  656     0
#> 221012_DGM_MB1599_13_2_432  496  264    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1590_5_2_402     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1597_11_1_425  256  109    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank1_1_1_390     0    0    0    0    0    0    0    0  856     0
#> 221012_DGM_MB1590_5_1_401     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1588_3_3_397     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1588_3_1_395     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank1_1_3_392     0    0    0    0    0    0    0    0  884     0
#> 221012_DGM_MB1592_7_2_411     0    0    0    0    0    0    0    0  185     0
#> 221012_DGM_MB1591_6_1_407     0    0    0    0    0    0    0    0  142     0
#> 221012_DGM_MB1589_4_1_398     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank4_1_3_436     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank3_1_3_421     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1592_7_3_412     0    0    0    0    0    0    0    0  194     0
#> 221012_DGM_MB1593_8_2_414     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1589_4_3_400     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1598_12_2_429    0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1589_4_2_399     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank2_1_2_405     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1591_6_2_408     0    0    0    0    0    0    0    0  126     0
#> 221012_DGM_Blank2_1_3_406     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1593_8_1_413     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1591_6_3_409     0    0    0    0    0    0    0    0  114     0
#> 221012_DGM_MB1592_7_1_410     0    0    0    0    0    0    0    0  147     0
#> 221012_DGM_MB1588_3_2_396     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1593_8_3_415     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1594_9_1_416     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1594_9_2_417     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_MB1594_9_3_418     0    0    0    0    0    0    0    0    0     0
#> 221012_DGM_Blank3_1_2_420     0    0    0    0    0    0    0    0    0     0
#>                            omu11 omu12 omu13 omu14 omu15 omu16 omu17 omu18
#> 221012_DGM_Blank4_1_2_435   2950     0     0     0     0     0     0     0
#> 221012_DGM_Blank4_1_1_434   2933     0     0     0     0     0     0     0
#> 221012_DGM_MB1599_13_3_433     0     0   161     0     0     0     0     0
#> 221012_DGM_MB1599_13_1_431     0     0   156     0     0     0     0     0
#> 221012_DGM_MB1598_12_3_430     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1598_12_1_428     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1597_11_2_426     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1595_10_3_424  2331     0     0     0     0     0     0     0
#> 221012_DGM_MB1595_10_2_423  1994     0     0     0     0     0     0     0
#> 221012_DGM_MB1597_11_3_427   101     0     0     0     0     0     0     0
#> 221012_DGM_MB1595_10_1_422  2289     0     0     0     0     0     0     0
#> 221012_DGM_Blank2_1_1_404   3324     0     0     0     0     0     0     0
#> 221012_DGM_Blank3_1_1_419      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1590_5_3_403   2081     0     0     0     0     0     0     0
#> 221012_DGM_Blank1_1_2_391      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1599_13_2_432     0     0   137     0     0     0     0     0
#> 221012_DGM_MB1590_5_2_402   2180     0     0     0     0     0     0     0
#> 221012_DGM_MB1597_11_1_425   113     0     0     0     0     0     0     0
#> 221012_DGM_Blank1_1_1_390      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1590_5_1_401   2215     0     0     0     0     0     0     0
#> 221012_DGM_MB1588_3_3_397      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1588_3_1_395      0     0     0     0     0     0     0     0
#> 221012_DGM_Blank1_1_3_392      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1592_7_2_411   2044     0     0     0     0     0     0     0
#> 221012_DGM_MB1591_6_1_407   1906     0     0     0     0     0     0     0
#> 221012_DGM_MB1589_4_1_398   2879     0     0     0     0     0     0     0
#> 221012_DGM_Blank4_1_3_436   3016     0     0     0     0     0     0     0
#> 221012_DGM_Blank3_1_3_421      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1592_7_3_412   2022     0     0     0     0     0     0     0
#> 221012_DGM_MB1593_8_2_414   1974     0     0     0     0     0     0     0
#> 221012_DGM_MB1589_4_3_400   2916     0     0     0     0     0     0     0
#> 221012_DGM_MB1598_12_2_429     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1589_4_2_399   2859     0     0     0     0     0     0     0
#> 221012_DGM_Blank2_1_2_405   3257     0     0     0     0     0     0     0
#> 221012_DGM_MB1591_6_2_408   2025     0     0     0     0     0     0     0
#> 221012_DGM_Blank2_1_3_406   3212     0     0     0     0     0     0     0
#> 221012_DGM_MB1593_8_1_413   2104     0     0     0     0     0     0     0
#> 221012_DGM_MB1591_6_3_409   2106     0     0     0     0     0     0     0
#> 221012_DGM_MB1592_7_1_410   1952     0     0     0     0     0     0     0
#> 221012_DGM_MB1588_3_2_396      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1593_8_3_415   1973     0     0     0     0     0     0     0
#> 221012_DGM_MB1594_9_1_416    723     0     0     0     0     0     0     0
#> 221012_DGM_MB1594_9_2_417    776     0     0     0     0     0     0     0
#> 221012_DGM_MB1594_9_3_418    763     0     0     0     0     0     0     0
#> 221012_DGM_Blank3_1_2_420      0     0     0     0     0     0     0     0
#>                            omu19 omu20 omu21 omu22 omu23 omu24 omu25 omu26
#> 221012_DGM_Blank4_1_2_435      0     0     0   176     0     0     0     0
#> 221012_DGM_Blank4_1_1_434      0     0     0   215     0     0     0     0
#> 221012_DGM_MB1599_13_3_433     0   268   256   554   513     0     0     0
#> 221012_DGM_MB1599_13_1_431     0   222   299   610   493     0     0     0
#> 221012_DGM_MB1598_12_3_430     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1598_12_1_428     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1597_11_2_426     0   179   208   197   151     0     0     0
#> 221012_DGM_MB1595_10_3_424     0     0     0   929   630     0     0     0
#> 221012_DGM_MB1595_10_2_423     0     0     0  1101   760     0     0     0
#> 221012_DGM_MB1597_11_3_427     0   163   192   205   151     0     0     0
#> 221012_DGM_MB1595_10_1_422     0     0     0   877   702     0     0     0
#> 221012_DGM_Blank2_1_1_404      0     0     0   120   154     0     0     0
#> 221012_DGM_Blank3_1_1_419      0     0     0     0  2765     0     0     0
#> 221012_DGM_MB1590_5_3_403      0     0     0  1016   887     0     0     0
#> 221012_DGM_Blank1_1_2_391      0     0     0  2219     0     0     0     0
#> 221012_DGM_MB1599_13_2_432     0   212   281   565   462     0     0     0
#> 221012_DGM_MB1590_5_2_402      0     0     0  1034   786     0     0     0
#> 221012_DGM_MB1597_11_1_425     0   173   209   182   153     0     0     0
#> 221012_DGM_Blank1_1_1_390      0     0     0  1909     0     0     0     0
#> 221012_DGM_MB1590_5_1_401      0     0     0   993   766     0     0     0
#> 221012_DGM_MB1588_3_3_397      0     0     0  2034  1688     0     0     0
#> 221012_DGM_MB1588_3_1_395      0     0     0  2002  1631     0     0     0
#> 221012_DGM_Blank1_1_3_392      0     0     0  1806     0     0     0     0
#> 221012_DGM_MB1592_7_2_411      0     0     0     0  1597     0     0     0
#> 221012_DGM_MB1591_6_1_407      0     0     0   918   747     0     0     0
#> 221012_DGM_MB1589_4_1_398      0     0     0   434   341     0     0     0
#> 221012_DGM_Blank4_1_3_436      0     0     0   169     0     0     0     0
#> 221012_DGM_Blank3_1_3_421      0     0     0     0  2796     0     0     0
#> 221012_DGM_MB1592_7_3_412      0     0     0     0  1618     0     0     0
#> 221012_DGM_MB1593_8_2_414      0     0     0  1008   762     0     0     0
#> 221012_DGM_MB1589_4_3_400      0     0     0   411   345     0     0     0
#> 221012_DGM_MB1598_12_2_429     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1589_4_2_399      0     0     0   462   359     0     0     0
#> 221012_DGM_Blank2_1_2_405      0     0     0   131   180     0     0     0
#> 221012_DGM_MB1591_6_2_408      0     0     0   946   766     0     0     0
#> 221012_DGM_Blank2_1_3_406      0     0     0   156   134     0     0     0
#> 221012_DGM_MB1593_8_1_413      0     0     0   960   698     0     0     0
#> 221012_DGM_MB1591_6_3_409      0     0     0   891   651     0     0     0
#> 221012_DGM_MB1592_7_1_410      0     0     0     0  1746     0     0     0
#> 221012_DGM_MB1588_3_2_396      0     0     0  2067  1574     0     0     0
#> 221012_DGM_MB1593_8_3_415      0     0     0  1045   734     0     0     0
#> 221012_DGM_MB1594_9_1_416      0     0     0  1126   727     0     0     0
#> 221012_DGM_MB1594_9_2_417      0     0     0   995   787     0     0     0
#> 221012_DGM_MB1594_9_3_418      0     0     0  1031   767     0     0     0
#> 221012_DGM_Blank3_1_2_420      0     0     0     0  2964     0     0     0
#>                            omu27 omu28 omu29 omu30 omu31 omu32 omu33 omu34
#> 221012_DGM_Blank4_1_2_435      0     0     0     0     0     0   874     0
#> 221012_DGM_Blank4_1_1_434      0     0     0     0     0     0   852     0
#> 221012_DGM_MB1599_13_3_433     0     0  1147     0     0     0     0     0
#> 221012_DGM_MB1599_13_1_431     0     0  1050     0     0     0     0     0
#> 221012_DGM_MB1598_12_3_430     0     0  3804   196     0     0     0     0
#> 221012_DGM_MB1598_12_1_428     0     0  3742   258     0     0     0     0
#> 221012_DGM_MB1597_11_2_426     0     0   780     0     0     0  1843     0
#> 221012_DGM_MB1595_10_3_424     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1595_10_2_423     0     0     0     0     0     0     0     0
#> 221012_DGM_MB1597_11_3_427     0     0   707     0     0     0  1916     0
#> 221012_DGM_MB1595_10_1_422     0     0     0     0     0     0     0     0
#> 221012_DGM_Blank2_1_1_404      0     0     0     0     0     0   402     0
#> 221012_DGM_Blank3_1_1_419   1235     0     0     0     0     0     0     0
#> 221012_DGM_MB1590_5_3_403      0     0     0     0     0     0   103     0
#> 221012_DGM_Blank1_1_2_391      0     0     0     0     0     0  1125     0
#> 221012_DGM_MB1599_13_2_432     0     0  1198     0     0     0     0     0
#> 221012_DGM_MB1590_5_2_402      0     0     0     0     0     0     0     0
#> 221012_DGM_MB1597_11_1_425     0     0   771     0     0     0  1810     0
#> 221012_DGM_Blank1_1_1_390      0     0     0     0     0     0  1235     0
#> 221012_DGM_MB1590_5_1_401      0     0     0     0     0     0   102     0
#> 221012_DGM_MB1588_3_3_397    278     0     0     0     0     0     0     0
#> 221012_DGM_MB1588_3_1_395    367     0     0     0     0     0     0     0
#> 221012_DGM_Blank1_1_3_392      0     0     0     0     0     0  1310     0
#> 221012_DGM_MB1592_7_2_411      0     0     0     0     0     0   174     0
#> 221012_DGM_MB1591_6_1_407      0     0     0     0     0     0   164   123
#> 221012_DGM_MB1589_4_1_398      0     0     0     0     0     0   202   144
#> 221012_DGM_Blank4_1_3_436      0     0     0     0     0     0   815     0
#> 221012_DGM_Blank3_1_3_421   1204     0     0     0     0     0     0     0
#> 221012_DGM_MB1592_7_3_412      0     0     0     0     0     0   166     0
#> 221012_DGM_MB1593_8_2_414      0     0     0     0     0     0   256     0
#> 221012_DGM_MB1589_4_3_400      0     0     0     0     0     0   214   114
#> 221012_DGM_MB1598_12_2_429     0     0  3769   231     0     0     0     0
#> 221012_DGM_MB1589_4_2_399      0     0     0     0     0     0   219   101
#> 221012_DGM_Blank2_1_2_405      0     0     0     0     0     0   432     0
#> 221012_DGM_MB1591_6_2_408      0     0     0     0     0     0   137     0
#> 221012_DGM_Blank2_1_3_406      0     0     0     0     0     0   498     0
#> 221012_DGM_MB1593_8_1_413      0     0     0     0     0     0   238     0
#> 221012_DGM_MB1591_6_3_409      0     0     0     0     0     0   126   112
#> 221012_DGM_MB1592_7_1_410      0     0     0     0     0     0   155     0
#> 221012_DGM_MB1588_3_2_396    359     0     0     0     0     0     0     0
#> 221012_DGM_MB1593_8_3_415      0     0     0     0     0     0   248     0
#> 221012_DGM_MB1594_9_1_416      0     0     0     0     0     0   479   945
#> 221012_DGM_MB1594_9_2_417      0     0     0     0     0     0   512   930
#> 221012_DGM_MB1594_9_3_418      0     0     0     0     0     0   482   957
#> 221012_DGM_Blank3_1_2_420   1036     0     0     0     0     0     0     0
#>                            omu35 omu36
#> 221012_DGM_Blank4_1_2_435      0     0
#> 221012_DGM_Blank4_1_1_434      0     0
#> 221012_DGM_MB1599_13_3_433   383     0
#> 221012_DGM_MB1599_13_1_431   397   105
#> 221012_DGM_MB1598_12_3_430     0     0
#> 221012_DGM_MB1598_12_1_428     0     0
#> 221012_DGM_MB1597_11_2_426   243     0
#> 221012_DGM_MB1595_10_3_424     0     0
#> 221012_DGM_MB1595_10_2_423     0     0
#> 221012_DGM_MB1597_11_3_427   205     0
#> 221012_DGM_MB1595_10_1_422     0     0
#> 221012_DGM_Blank2_1_1_404      0     0
#> 221012_DGM_Blank3_1_1_419      0     0
#> 221012_DGM_MB1590_5_3_403      0     0
#> 221012_DGM_Blank1_1_2_391      0     0
#> 221012_DGM_MB1599_13_2_432   385     0
#> 221012_DGM_MB1590_5_2_402      0     0
#> 221012_DGM_MB1597_11_1_425   224     0
#> 221012_DGM_Blank1_1_1_390      0     0
#> 221012_DGM_MB1590_5_1_401      0     0
#> 221012_DGM_MB1588_3_3_397      0     0
#> 221012_DGM_MB1588_3_1_395      0     0
#> 221012_DGM_Blank1_1_3_392      0     0
#> 221012_DGM_MB1592_7_2_411      0     0
#> 221012_DGM_MB1591_6_1_407      0     0
#> 221012_DGM_MB1589_4_1_398      0     0
#> 221012_DGM_Blank4_1_3_436      0     0
#> 221012_DGM_Blank3_1_3_421      0     0
#> 221012_DGM_MB1592_7_3_412      0     0
#> 221012_DGM_MB1593_8_2_414      0     0
#> 221012_DGM_MB1589_4_3_400      0     0
#> 221012_DGM_MB1598_12_2_429     0     0
#> 221012_DGM_MB1589_4_2_399      0     0
#> 221012_DGM_Blank2_1_2_405      0     0
#> 221012_DGM_MB1591_6_2_408      0     0
#> 221012_DGM_Blank2_1_3_406      0     0
#> 221012_DGM_MB1593_8_1_413      0     0
#> 221012_DGM_MB1591_6_3_409      0     0
#> 221012_DGM_MB1592_7_1_410      0     0
#> 221012_DGM_MB1588_3_2_396      0     0
#> 221012_DGM_MB1593_8_3_415      0     0
#> 221012_DGM_MB1594_9_1_416      0     0
#> 221012_DGM_MB1594_9_2_417      0     0
#> 221012_DGM_MB1594_9_3_418      0     0
#> 221012_DGM_Blank3_1_2_420      0     0